Torch.Indef.Dynamic.NN

Description

NN backpack API which is _not recommended to be used directly_. This package is just a wrapper around the backpack signatures of Torch's C-based NN library.

Instead, use NN which includes an AD abstraction, simple forward- and backward- functions, and staticly-typed dimension-level checking.

This library will, over time, iterate into type-safe code -- but for the moment we hoist the safety into the dependent types, which is faster to iterate with and is semantically clearer for development (the errors take a bit of getting used to).

Documentation

module Torch.Indef.Dynamic.NN.Pooling

module Torch.Indef.Dynamic.NN.Activation

_abs_updateOutput :: Dynamic -> Dynamic -> IO ()

abs forward

_abs_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

abs backward

_absCriterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

absCriterion forward pass (updates the output tensor)

_absCriterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

absCriterion backward-update (updates the layer and bias tensors)

_bCECriterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Dynamic -> Bool -> IO ()

bCECriterion forward pass (updates the output tensor)

_bCECriterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Dynamic -> Bool -> IO ()

bCECriterion backward-update (updates the layer and bias tensors)

_distKLDivCriterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

distKLDivCriterion forward pass (updates the output tensor)

_distKLDivCriterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

distKLDivCriterion backward-update (updates the layer and bias tensors)

_gatedLinear_updateOutput :: Dynamic -> Dynamic -> Int -> IO ()

gatedLinear forward pass (updates the output tensor)

_gatedLinear_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

gatedLinear backward-update (updates the layer and bias tensors)

_hardTanh_updateOutput :: Dynamic -> Dynamic -> Double -> Double -> Bool -> IO ()

hardTanh forward pass (updates the output tensor)

_hardTanh_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Double -> Double -> Bool -> IO ()

hardTanh backward-update (updates the layer and bias tensors)

_im2Col_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

im2Col forward pass (updates the output tensor)

_im2Col_updateGradInput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

im2Col backward-update (updates the layer and bias tensors)

_col2Im_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

col2Im forward pass (updates the output tensor)

_col2Im_updateGradInput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

col2Im backward-update (updates the layer and bias tensors)

_l1Cost_updateOutput :: Dynamic -> Dynamic -> IO ()

l1Cost forward pass (updates the output tensor)

_l1Cost_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

l1Cost backward-update (updates the layer and bias tensors)

_gRUFused_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

gRUFused forward pass (updates the output tensor)

_gRUFused_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

gRUFused backward-update (updates the layer and bias tensors)

_lSTMFused_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

lSTMFused forward pass (updates the output tensor)

_lSTMFused_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

lSTMFused backward-update (updates the layer and bias tensors)

_logSigmoid_updateOutput :: Dynamic -> Dynamic -> Dynamic -> IO ()

logSigmoid forward pass (updates the output tensor)

_logSigmoid_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

logSigmoid backward-update (updates the layer and bias tensors)

_logSoftMax_updateOutput :: Dynamic -> Dynamic -> Integer -> IO ()

logSoftMax forward pass (updates the output tensor)

_logSoftMax_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Integer -> IO ()

logSoftMax backward-update (updates the layer and bias tensors)

_marginCriterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Double -> IO ()

marginCriterion forward pass (updates the output tensor)

_marginCriterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Double -> IO ()

marginCriterion backward-update (updates the layer and bias tensors)

_softMarginCriterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

softMarginCriterion forward pass (updates the output tensor)

_softMarginCriterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

softMarginCriterion backward-update (updates the layer and bias tensors)

_mSECriterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

mSECriterion forward pass (updates the output tensor)

_mSECriterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

mSECriterion backward-update (updates the layer and bias tensors)

_sigmoid_updateOutput :: Dynamic -> Dynamic -> IO ()

sigmoid forward pass (updates the output tensor)

_sigmoid_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

sigmoid backward-update (updates the layer and bias tensors)

_smoothL1Criterion_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

smoothL1Criterion forward pass (updates the output tensor)

_smoothL1Criterion_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Bool -> IO ()

smoothL1Criterion backward-update (updates the layer and bias tensors)

_softMax_updateOutput :: Dynamic -> Dynamic -> Integer -> IO ()

softMax forward pass (updates the output tensor)

_softMax_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Integer -> IO ()

softMax backward-update (updates the layer and bias tensors)

_softPlus_updateOutput :: Dynamic -> Dynamic -> Double -> Double -> IO ()

softPlus forward pass (updates the output tensor)

_softPlus_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> Double -> IO ()

softPlus backward-update (updates the layer and bias tensors)

_softShrink_updateOutput :: Dynamic -> Dynamic -> Double -> IO ()

softShrink forward pass (updates the output tensor)

_softShrink_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Double -> IO ()

softShrink backward-update (updates the layer and bias tensors)

_sparseLinear_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

sparseLinear forward pass (updates the output tensor)

_sparseLinear_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> Double -> IO ()

sparseLinear backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_sparseLinear_zeroGradParameters :: Dynamic -> Dynamic -> Dynamic -> IO ()

sparseLinear zeroGradParameters

_sparseLinear_updateParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> IO ()

sparseLinear updateParameters

_sparseLinear_legacyUpdateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

sparseLinear legacyUpdateOutput

_sparseLinear_legacyAccGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> Double -> IO ()

sparseLinear legacyAccGradParameters

_sqrt_updateOutput :: Dynamic -> Dynamic -> Double -> IO ()

sqrt forward pass (updates the output tensor)

_sqrt_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

sqrt backward-update (updates the layer and bias tensors)

_square_updateOutput :: Dynamic -> Dynamic -> IO ()

square forward pass (updates the output tensor)

_square_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

square backward-update (updates the layer and bias tensors)

_tanh_updateOutput :: Dynamic -> Dynamic -> IO ()

tanh forward pass (updates the output tensor)

_tanh_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

tanh backward-update (updates the layer and bias tensors)

_temporalConvolution_updateOutput

Arguments

:: Dynamic
input

-> Dynamic
output -- this is the mutated return value

-> Dynamic
2d weight tensor

-> Dynamic
1d bias tensor

-> Int
kernel width

-> Int
step size

-> Int
feature size

-> Int
output size

-> IO ()

temporalConvolution forward pass (updates the output tensor)

_temporalConvolution_updateGradInput

Arguments

:: Dynamic
input

-> Dynamic
grad output

-> Dynamic
grad input -- this is the mutated return value

-> Dynamic
weights

-> Int
kernel width

-> Int
step size

-> IO ()

temporalConvolution backward-update (updates the layer and bias tensors)

_temporalConvolution_accGradParameters

Arguments

:: Dynamic
input

-> Dynamic
grad output

-> Dynamic
grad weight -- this is a mutated argument and torch will accumulate this gradient

-> Dynamic
grad bias -- this is a mutated argument and torch will accumulate this gradient

-> Int
kernel width

-> Int
step size

-> Double
scale

-> IO ()

temporalConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_temporalRowConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Bool -> IO ()

temporalRowConvolution forward pass (updates the output tensor)

_temporalRowConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Bool -> IO ()

temporalRowConvolution backward-update (updates the layer and bias tensors)

_temporalRowConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Bool -> Double -> IO ()

temporalRowConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_temporalUpSamplingNearest_updateOutput :: Dynamic -> Dynamic -> Int -> IO ()

temporalUpSamplingNearest forward pass (updates the output tensor)

_temporalUpSamplingNearest_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

temporalUpSamplingNearest backward-update (updates the layer and bias tensors)

_temporalUpSamplingLinear_updateOutput :: Dynamic -> Dynamic -> Int -> IO ()

temporalUpSamplingLinear forward pass (updates the output tensor)

_temporalUpSamplingLinear_updateGradInput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

temporalUpSamplingLinear backward-update (updates the layer and bias tensors)

_batchNormalization_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Double -> Double -> IO ()

batchNormalization forward pass (updates the output tensor)

_batchNormalization_backward :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Bool -> Double -> Double -> IO ()

batchNormalization backward

spatialConvolutionMM_updateOutput

Arguments

:: Dynamic
input

-> Dynamic
3D weight tensor (connTable:size(1) x kH x kW)

-> Dynamic
1D bias tensor (nOutputPlane)

-> (Int, Int)
(kW, kH) kernel height and width

-> (Int, Int)
(dW, dH) step of the convolution in width and height dimensions. C-default is 1 for both.

-> (Int, Int)
(pW, pH) zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> IO Dynamic
output

spatialConvolutionMM forward pass

_spatialConvolutionMM_updateGradInput

Arguments

:: Dynamic
input

-> Dynamic
gradOutput

-> Dynamic
gradInput

-> Dynamic
weight

-> Dynamic
columns

-> Dynamic
ones

-> (Int, Int)
(kW, kH) kernel height and width

-> (Int, Int)
(dW, dH) step of the convolution in width and height dimensions

-> (Int, Int)
(pW, pH) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> IO ()

spatialConvolutionMM backward-update (updates the layer and bias tensors)

_spatialConvolutionMM_accGradParameters

Arguments

:: Dynamic
input

-> Dynamic
gradOutput

-> Dynamic
gradWeight

-> Dynamic
gradBias

-> Dynamic
finput/columns <<- required. This can be NULL in C if gradWeight is NULL.

-> Dynamic
ones

-> (Int, Int)
(kW, kH) kernel height and width

-> (Int, Int)
(dW, dH) step of the convolution in width and height dimensions

-> (Int, Int)
(pW, pH) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double

-> IO ()

spatialConvolutionMM backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialConvolutionLocal_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> CLLong -> CLLong -> CLLong -> CLLong -> IO ()

spatialConvolutionLocal forward pass (updates the output tensor)

_spatialConvolutionLocal_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> CLLong -> CLLong -> CLLong -> CLLong -> IO ()

spatialConvolutionLocal backward-update (updates the layer and bias tensors)

_spatialConvolutionLocal_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> CLLong -> CLLong -> CLLong -> CLLong -> Double -> IO ()

spatialConvolutionLocal backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialFullConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialFullConvolution forward pass (updates the output tensor)

_spatialFullConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialFullConvolution backward-update (updates the layer and bias tensors)

_spatialFullConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

spatialFullConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialDilatedConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialDilatedConvolution forward pass (updates the output tensor)

_spatialDilatedConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialDilatedConvolution backward-update (updates the layer and bias tensors)

_spatialDilatedConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

spatialDilatedConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialFullDilatedConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialFullDilatedConvolution forward pass (updates the output tensor)

_spatialFullDilatedConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialFullDilatedConvolution backward-update (updates the layer and bias tensors)

_spatialFullDilatedConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

spatialFullDilatedConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialSubSampling_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

spatialSubSampling forward pass (updates the output tensor)

_spatialSubSampling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

spatialSubSampling backward-update (updates the layer and bias tensors)

_spatialSubSampling_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Double -> IO ()

spatialSubSampling backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialUpSamplingNearest_updateOutput :: Dynamic -> Dynamic -> Int -> IO ()

spatialUpSamplingNearest forward pass (updates the output tensor)

_spatialUpSamplingNearest_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

spatialUpSamplingNearest backward-update (updates the layer and bias tensors)

_spatialUpSamplingBilinear_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> IO ()

spatialUpSamplingBilinear forward pass (updates the output tensor)

_spatialUpSamplingBilinear_updateGradInput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialUpSamplingBilinear backward-update (updates the layer and bias tensors)

_spatialGridSamplerBilinear_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

spatialGridSamplerBilinear forward pass (updates the output tensor)

_spatialGridSamplerBilinear_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

spatialGridSamplerBilinear backward-update (updates the layer and bias tensors)

_volumetricGridSamplerBilinear_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

volumetricGridSamplerBilinear forward pass (updates the output tensor)

_volumetricGridSamplerBilinear_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

volumetricGridSamplerBilinear backward-update (updates the layer and bias tensors)

_volumetricConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricConvolution forward pass (updates the output tensor)

_volumetricConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricConvolution backward-update (updates the layer and bias tensors)

_volumetricConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

volumetricConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_volumetricFullConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricFullConvolution forward pass (updates the output tensor)

_volumetricFullConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricFullConvolution backward-update (updates the layer and bias tensors)

_volumetricFullConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

volumetricFullConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_volumetricDilatedConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricDilatedConvolution forward pass (updates the output tensor)

_volumetricDilatedConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricDilatedConvolution backward-update (updates the layer and bias tensors)

_volumetricDilatedConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

volumetricDilatedConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_volumetricFullDilatedConvolution_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricFullDilatedConvolution forward pass (updates the output tensor)

_volumetricFullDilatedConvolution_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricFullDilatedConvolution backward-update (updates the layer and bias tensors)

_volumetricFullDilatedConvolution_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Double -> IO ()

volumetricFullDilatedConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialReflectionPadding_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

spatialReflectionPadding forward pass (updates the output tensor)

_spatialReflectionPadding_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

spatialReflectionPadding backward-update (updates the layer and bias tensors)

_spatialReplicationPadding_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

spatialReplicationPadding forward pass (updates the output tensor)

_spatialReplicationPadding_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IO ()

spatialReplicationPadding backward-update (updates the layer and bias tensors)

_volumetricReplicationPadding_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricReplicationPadding forward pass (updates the output tensor)

_volumetricReplicationPadding_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricReplicationPadding backward-update (updates the layer and bias tensors)

_volumetricUpSamplingNearest_updateOutput :: Dynamic -> Dynamic -> Int -> IO ()

volumetricUpSamplingNearest forward pass (updates the output tensor)

_volumetricUpSamplingNearest_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> IO ()

volumetricUpSamplingNearest backward-update (updates the layer and bias tensors)

_volumetricUpSamplingTrilinear_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> IO ()

volumetricUpSamplingTrilinear forward pass (updates the output tensor)

_volumetricUpSamplingTrilinear_updateGradInput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricUpSamplingTrilinear backward-update (updates the layer and bias tensors)

_temporalReflectionPadding_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> IO ()

temporalReflectionPadding forward pass (updates the output tensor)

_temporalReflectionPadding_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> IO ()

temporalReflectionPadding backward-update (updates the layer and bias tensors)

_temporalReplicationPadding_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> IO ()

temporalReplicationPadding forward pass (updates the output tensor)

_temporalReplicationPadding_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> IO ()

temporalReplicationPadding backward-update (updates the layer and bias tensors)