Torch.Indef.Dynamic.NN.Activation

Description

DYNAMIC-NN MODULE WARNING: this module is mostly unfinished and undocumented. It provides, in essence, direct calls to the torch neural network libraries: THNN and THCUNN. Because the dynamic tensor code requires a lot of runtime checks which requires a lot of thought regarding a good API, the recommended route is to use Static tensors, which have a much more natural API and is inherently safer.

Documentation

_threshold_updateOutput :: Dynamic -> Dynamic -> Double -> Double -> Bool -> IO ()

threshold forward pass (updates the output tensor)

_threshold_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Double -> Double -> Bool -> IO ()

threshold backward-update (updates the layer and bias tensors)

_pReLU_updateOutput :: Dynamic -> Dynamic -> Dynamic -> IO ()

pReLU forward pass (updates the output tensor)

_pReLU_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> IO ()

pReLU backward-update (updates the layer and bias tensors)

_pReLU_accGradParameters :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> IO ()

pReLU backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_rReLU_updateOutput :: Dynamic -> Dynamic -> Dynamic -> Double -> Double -> Bool -> Bool -> Generator -> IO ()

rReLU forward pass (updates the output tensor)

_rReLU_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> Double -> Bool -> Bool -> IO ()

rReLU backward-update (updates the layer and bias tensors)

_leakyReLU_updateOutput :: Dynamic -> Dynamic -> Double -> Bool -> IO ()

leakyReLU forward pass (updates the output tensor)

_leakyReLU_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Double -> Bool -> IO ()

leakyReLU backward-update (updates the layer and bias tensors)

_eLU_updateOutput :: Dynamic -> Dynamic -> Double -> Double -> Bool -> IO ()

eLU forward pass (updates the output tensor)

_eLU_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Double -> Double -> IO ()

eLU backward-update (updates the layer and bias tensors)