Torch.Indef.Dynamic.NN.Pooling

Description

DYNAMIC-NN MODULE WARNING: this module is mostly unfinished and undocumented. It provides, in essence, direct calls to the torch neural network libraries: THNN and THCUNN. Because the dynamic tensor code requires a lot of runtime checks which requires a lot of thought regarding a good API, the recommended route is to use Static tensors, which have a much more natural API and is inherently safer.

Documentation

_spatialAdaptiveAveragePooling_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> IO ()

spatialAdaptiveAveragePooling forward pass (updates the output tensor)

_spatialAdaptiveAveragePooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

spatialAdaptiveAveragePooling backward pass (updates the gradInput tensor)

_spatialAveragePooling_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

spatialAveragePooling forward pass (updates the output tensor)

_spatialAveragePooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

spatialAveragePooling backward pass (updates the gradInput tensor)

_volumetricAveragePooling_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

volumetricAveragePooling forward pass (updates the output tensor)

_volumetricAveragePooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

volumetricAveragePooling backward pass (updates the gradInput tensor)

_volumetricAdaptiveAveragePooling_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> IO ()

volumetricAdaptiveAveragePooling forward pass (updates the output tensor)

_volumetricAdaptiveAveragePooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IO ()

volumetricAdaptiveAveragePooling backward pass (updates the gradInput tensor)

_featureLPPooling_updateOutput :: Dynamic -> Dynamic -> Double -> Int -> Int -> Bool -> IO ()

featureLPPooling forward pass (updates the output tensor)

_featureLPPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Dynamic -> Double -> Int -> Int -> Bool -> IO ()

featureLPPooling backward pass (updates the gradInput tensor)

_temporalMaxPooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> IO ()

temporalMaxPooling forward pass (updates the output tensor)

_temporalMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> IO ()

temporalMaxPooling backward pass (updates the gradInput tensor)

_spatialAdaptiveMaxPooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> IO ()

spatialAdaptiveMaxPooling forward pass (updates the output tensor)

_spatialAdaptiveMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> IO ()

spatialAdaptiveMaxPooling backward pass (updates the gradInput tensor)

_spatialFractionalMaxPooling_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IndexDynamic -> Dynamic -> IO ()

spatialFractionalMaxPooling forward pass (updates the output tensor)

_spatialFractionalMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> IndexDynamic -> IO ()

spatialFractionalMaxPooling backward pass (updates the gradInput tensor)

_spatialMaxPooling_updateOutput

Arguments

:: Dynamic
input

-> Dynamic
output

-> IndexDynamic
indices

-> (Int, Int)
kernel size

-> (Int, Int)
step size

-> (Int, Int)
padding size

-> Bool
ceil mode

-> IO ()

spatialMaxPooling forward pass (updates the output tensor)

_spatialMaxPooling_updateGradInput

Arguments

:: Dynamic
input

-> Dynamic
grad output

-> Dynamic
grad input

-> IndexDynamic
indices

-> (Int, Int)
kernel size

-> (Int, Int)
step size

-> (Int, Int)
padding size

-> Bool
ceil mode

-> IO ()

spatialMaxPooling backward pass (updates the gradInput tensor)

_spatialDilatedMaxPooling_updateOutput

Arguments

:: Dynamic
input

-> Dynamic
output

-> IndexDynamic
indices

-> (Int, Int)
kernel size

-> (Int, Int)
step size

-> (Int, Int)
padding size

-> (Int, Int)
dilation size

-> Bool
ceil mode

-> IO ()

spatialDilatedMaxPooling forward pass (updates the output tensor)

_spatialDilatedMaxPooling_updateGradInput

Arguments

:: Dynamic
input

-> Dynamic
grad output

-> Dynamic
grad input

-> IndexDynamic
indices

-> (Int, Int)
kernel size

-> (Int, Int)
step size

-> (Int, Int)
padding size

-> (Int, Int)
dilation size

-> Bool
ceil mode

-> IO ()

spatialDilatedMaxPooling backward pass (updates the gradInput tensor)

_spatialMaxUnpooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> IO ()

spatialMaxUnpooling forward pass (updates the output tensor)

_spatialMaxUnpooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> IO ()

spatialMaxUnpooling backward pass (updates the gradInput tensor)

_volumetricFractionalMaxPooling_updateOutput :: Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IndexDynamic -> Dynamic -> IO ()

volumetricFractionalMaxPooling forward pass (updates the output tensor)

_volumetricFractionalMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> Int -> Int -> Int -> Int -> Int -> Int -> IndexDynamic -> IO ()

volumetricFractionalMaxPooling backward pass (updates the gradInput tensor)

_volumetricMaxPooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricMaxPooling forward pass (updates the output tensor)

_volumetricMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricMaxPooling backward pass (updates the gradInput tensor)

_volumetricDilatedMaxPooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricDilatedMaxPooling forward pass (updates the output tensor)

_volumetricDilatedMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricDilatedMaxPooling backward pass (updates the gradInput tensor)

_volumetricMaxUnpooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricMaxUnpooling forward pass (updates the output tensor)

_volumetricMaxUnpooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricMaxUnpooling backward pass (updates the gradInput tensor)

_volumetricAdaptiveMaxPooling_updateOutput :: Dynamic -> Dynamic -> IndexDynamic -> Int -> Int -> Int -> IO ()

volumetricAdaptiveMaxPooling forward pass (updates the output tensor)

_volumetricAdaptiveMaxPooling_updateGradInput :: Dynamic -> Dynamic -> Dynamic -> IndexDynamic -> IO ()

volumetricAdaptiveMaxPooling backward pass (updates the gradInput tensor)