Torch.Indef.Dynamic.Tensor

Description

This package is the class for handling numeric data in dynamic tensors.

A Dynamic is a multi-dimensional matrix without static type-level dimensions. The number of dimensions is unlimited (up to what can be created using LongStorage).

Documentation

_clearFlag :: Dynamic -> Int8 -> IO ()

Clears the internal flags on a tensor. Uses bitwise operators for flags.

tensordata :: Dynamic -> IO [HsReal]

get the underlying data as a haskell list from the tensor

get1d :: Dynamic -> Int64 -> HsReal

get a value from dimension 1

get2d :: Dynamic -> Int64 -> Int64 -> HsReal

get a value from dimension 2

get3d :: Dynamic -> Int64 -> Int64 -> Int64 -> HsReal

get a value from dimension 3

get4d :: Dynamic -> Int64 -> Int64 -> Int64 -> Int64 -> HsReal

get a value from dimension 4

isContiguous :: Dynamic -> IO Bool

whether or not the tensor is contiguous in memory.

isSameSizeAs :: Dynamic -> Dynamic -> IO Bool

check to see if to tensors are the same size as eachother.

isSetTo :: Dynamic -> Dynamic -> IO Bool

Returns true iff the Tensor is set to the argument Tensor.

Note: this is only true if the tensors are the same size, have the same strides and share the same storage and offset.

isSize :: Dynamic -> LongStorage -> IO Bool

check to see if the tensor is the same size as the LongStorage.

nDimension :: Dynamic -> IO Int32

Returns the number of dimensions in a Tensor.

nElement :: Dynamic -> IO Int64

Returns the number of elements in a Tensor.

_narrow :: Dynamic -> Dynamic -> DimVal -> Int64 -> Size -> IO ()

Warning: hasktorch devs have not yet made this safe. You are warned.

returns a tensor which shares the same Storage as the original. Hence, any modification in the memory of the sub-tensor will have an impact on the primary tensor, and vice-versa. These methods are very fast, as they do not involve any memory copy.

empty :: IO Dynamic

Returns an empty tensor.

newExpand :: Dynamic -> IndexStorage -> IO Dynamic

pure version of _expand

_expand

Arguments

:: Dynamic
return tensor to mutate inplace.

-> Dynamic
source tensor to expand

-> IndexStorage
how to expand the tensor.

-> IO ()

Expanding a tensor does not allocate new memory, but only creates a new view on the existing tensor where singleton dimensions can be expanded to multiple ones by setting the stride to 0. Any dimension that has size 1 can be expanded to arbitrary value without any new memory allocation. Attempting to expand along a dimension that does not have size 1 will result in an error which we do not currently handle in hasktorch.

_expandNd :: NonEmpty Dynamic -> NonEmpty Dynamic -> Int -> IO ()

FIXME: doublecheck what this does.

newClone :: Dynamic -> IO Dynamic

purely clone a tensor

newContiguous :: Dynamic -> IO Dynamic

purely clone a tensor to have a contiguous memory layout.

newNarrow :: Dynamic -> DimVal -> Int64 -> Size -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

returns a tensor which shares the same Storage as the original. Hence, any modification in the memory of the sub-tensor will have an impact on the primary tensor, and vice-versa. These methods are very fast, as they do not involve any memory copy.

newSelect :: Dynamic -> DimVal -> Int64 -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

returns a tensor which shares the same Storage as the original. Hence, any modification in the memory of the sub-tensor will have an impact on the primary tensor, and vice-versa. These methods are very fast, as they do not involve any memory copy.

newSizeOf :: Dynamic -> IO IndexStorage

get the sizes of each dimension

FIXME: doublecheck this

newStrideOf :: Dynamic -> IO IndexStorage

get the strides of each dimension

FIXME: doublecheck this

newTranspose :: Dynamic -> DimVal -> DimVal -> IO Dynamic

pure version of _transpose

newUnfold :: Dynamic -> DimVal -> Int64 -> Int64 -> IO Dynamic

pure version of _unfold

newView :: Dynamic -> IndexStorage -> IO Dynamic

Creates a view with different dimensions of the storage associated with tensor, returning a new tensor.

FIXME: I think resizeAs is the non-cloning version of this function. See: https://github.com/torch/torch7/blob/master/doc/tensor.md#result-viewresult-tensor-sizes

for more.

newWithSize :: IndexStorage -> IndexStorage -> IO Dynamic

create an uninitialized tensor with the given size and strides (?)

FIXME: doublecheck what the IndexStorages stands for

newWithSize1d :: Size -> IO Dynamic

create an uninitialized 1d tensor

newWithSize2d :: Size -> Size -> IO Dynamic

create an uninitialized 2d tensor

newWithSize3d :: Size -> Size -> Size -> IO Dynamic

create an uninitialized 3d tensor

newWithSize4d :: Size -> Size -> Size -> Size -> IO Dynamic

create an uninitialized 4d tensor

newWithStorage :: Storage -> StorageOffset -> IndexStorage -> IndexStorage -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

create a new tensor with the given size and strides, storage offset and storage.

FIXME: doublecheck what all of this does.

newWithStorage1d :: Storage -> StorageOffset -> (Size, Stride) -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

create a new 1d tensor with the given storage's first dimension.

newWithStorage2d :: Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

create a new 2d tensor with the given storage's first 2 dimensions.

newWithStorage3d :: Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

create a new 3d tensor with the given storage's first 3 dimensions.

newWithStorage4d :: Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

create a new 4d tensor with the given storage's first 4 dimensions.

newWithTensor :: Dynamic -> IO Dynamic

Warning: hasktorch devs have not yet made this safe. You are warned.

create a new tensor with the given tensor's underlying storage.

_resize :: Dynamic -> IndexStorage -> IndexStorage -> IO ()

Resize the tensor according to the given LongStorage size (and strides?) FIXME: doublecheck what the IndexStorages stands for

_resize1d :: Dynamic -> Int64 -> IO ()

resize dimension 1 of a tensor.

_resize2d :: Dynamic -> Int64 -> Int64 -> IO ()

resize the first 2 dimensions of a tensor.

_resize3d :: Dynamic -> Int64 -> Int64 -> Int64 -> IO ()

resize the first 3 dimensions of a tensor.

_resize4d :: Dynamic -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()

resize the first 4 dimensions of a tensor.

_resize5d :: Dynamic -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO ()

resize the first 5 dimensions of a tensor.

_resizeAs :: Dynamic -> Dynamic -> IO ()

Resize the tensor as the given tensor.

_resizeNd

Arguments

:: Dynamic
tensor to resize inplace.

-> Int32
unknown argument. FIXME: Someone needs to find this out.

-> [Size]
new sizes to update

-> [Stride]
new strides to update.

-> IO ()

resize a tensor with given strides, sizes and a magical parameter.

FIXME: Someone needs to find out what the magical parameter is.

retain :: Dynamic -> IO ()

Increment the reference counter of the tensor.

From: https://github.com/torch/torch7/blob/aed31711c6b8846b8337a263a7f9f998697994e7/doc/tensor.md#reference-counting

Tensors are reference-counted. It means that each time an object (C or the Lua state) need to keep a reference over a tensor, the corresponding tensor reference counter will be increased. The reference counter is decreased when the object does not need the tensor anymore.

These methods should be used with extreme care. In general, they should never be called, except if you know what you are doing, as the handling of references is done automatically. They can be useful in threaded environments. Note that these methods are atomic operations.

_select :: Dynamic -> Dynamic -> DimVal -> Int64 -> IO ()

Warning: hasktorch devs have not yet made this safe. You are warned.

returns a tensor which shares the same Storage as the original. Hence, any modification in the memory of the sub-tensor will have an impact on the primary tensor, and vice-versa. These methods are very fast, as they do not involve any memory copy.

_set

Arguments

:: Dynamic
the source tensor which is mutated inplace

-> Dynamic
the tensor who's storage is going to be referenced.

-> IO ()

Warning: hasktorch devs have not yet made this safe. You are warned.

set the source tensor's storage to another tensor.

_set1d :: Dynamic -> Int64 -> HsReal -> IO ()

set a value in dimension 1, inplace.

_set2d :: Dynamic -> Int64 -> Int64 -> HsReal -> IO ()

set a value in dimension 2, inplace.

_set3d :: Dynamic -> Int64 -> Int64 -> Int64 -> HsReal -> IO ()

set a value in dimension 3, inplace.

_set4d :: Dynamic -> Int64 -> Int64 -> Int64 -> Int64 -> HsReal -> IO ()

set a value in dimension 4, inplace.

_setFlag :: Dynamic -> Int8 -> IO ()

set the flags on a tensor inplace

_setStorage :: Dynamic -> Storage -> StorageOffset -> IndexStorage -> IndexStorage -> IO ()

Set the storage of a tensor.

FIXME: doublecheck what the IndexStorages stands for

_setStorage1d :: Dynamic -> Storage -> StorageOffset -> (Size, Stride) -> IO ()

Set the storage of a tensor, only referencing 1 dimension of storage

FIXME: find out what storageoffset does.

_setStorage2d :: Dynamic -> Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> IO ()

Set the storage of a tensor, only referencing 2 dimensions of storage

FIXME: find out what storageoffset does.

_setStorage3d :: Dynamic -> Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO ()

Set the storage of a tensor, only referencing 3 dimensions of storage

FIXME: find out what storageoffset does.

_setStorage4d :: Dynamic -> Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO ()

Set the storage of a tensor, only referencing 4 dimensions of storage

FIXME: find out what storageoffset does.

_setStorageNd :: Dynamic -> Storage -> StorageOffset -> DimVal -> [Size] -> [Stride] -> IO ()

Set the storage of a tensor, referencing any number of dimensions of storage

size :: Dynamic -> DimVal -> IO Size

get the size of a tensor's specific dimension.

sizeDesc :: Dynamic -> IO DescBuff

primarily used for debugging. Get the size description from a c call.

_squeeze :: Dynamic -> Dynamic -> IO ()

Removes all singleton dimensions of the tensor.

_squeeze1d :: Dynamic -> Dynamic -> DimVal -> IO ()

Removes a singleton dimensions of the tensor at a given dimension.

storage :: Dynamic -> IO Storage

get the underlying storage of a tensor

storageOffset :: Dynamic -> IO StorageOffset

get the storage offset of a tensor

stride :: Dynamic -> DimVal -> IO Stride

Returns the jump necessary to go from one element to the next one in the specified dimension dim.

_transpose

Arguments

:: Dynamic
tensor to mutate into the result.

-> Dynamic
source tensor to use for data.

-> DimVal
dim1

-> DimVal
dim2

-> IO ()

Returns a tensor where dimensions dim1 and dim2 have been swapped.

_unfold

Arguments

:: Dynamic
tensor to mutate into the result.

-> Dynamic
source tensor to use for data.

-> DimVal

-> Size

-> Step

-> IO ()

Returns a tensor which contains all slices of size size in the dimension dim. Step between two slices is given by step.

If sizedim is the original size of dimension dim, the size of dimension dim in the returned tensor will be (sizedim - size) / step + 1

An additional dimension of size size is appended in the returned tensor.

FIXME: this still takes C-like arguments which mutates the first arg inplace.

_unsqueeze1d

Arguments

:: Dynamic
tensor to mutate into the result.

-> Dynamic
source tensor to use for data.

-> DimVal
dimension to unsqueeze

-> IO ()

unsqueeze a tensor, adding a singleton dimension at the specified dimval.

shape :: Dynamic -> IO [Size]

return the a runtime shape representing the dimensions of a Dynamic

_setStorageDim :: Dynamic -> Storage -> StorageOffset -> [(Size, Stride)] -> IO ()

set the storage dimensionality of a dynamic tensor, inplace, to any new size and stride pair.

_setDim :: Dynamic -> Dims (d :: [Nat]) -> HsReal -> IO ()

set a value of a dynamic tensor, inplace, with any dimensionality.

_resizeDim :: Dynamic -> Dims (d :: [Nat]) -> IO ()

resize a dynamic tensor, inplace, to any new dimensionality

vector :: [HsReal] -> Dynamic

create a 1d Dynamic tensor from a list of elements.

FIXME construct this with TH, not by using setDim inplace (one-by-one) which might be doing a second linear pass.

resizeDim :: Dynamic -> Dims (d :: [Nat]) -> IO Dynamic

resize a dynamic tensor with runtime SomeDims representation of its new shape. Returns a pure copy of the input tensor.

resizeDim' :: Dynamic -> SomeDims -> IO Dynamic

alias to resizeDim with a runtime SomeDims

getDim :: Dynamic -> Dims (d :: [Nat]) -> IO HsReal

get a value from a dynamic tensor at a given index, Dims d.

getDims :: Dynamic -> IO SomeDims

alias to getDimList which wraps the dimensions list in a SomeDims

getDimList :: Integral i => Dynamic -> IO [i]

get the runtime dimension list of a dynamic tensor

new :: forall (d :: [Nat]). Dims d -> IO Dynamic

create a new dynamic tensor of size Dims d

setDim'_ :: Dynamic -> SomeDims -> HsReal -> IO ()

set a specific runtime SomeDims position of a dynamic tensor.

resizeDim'_ :: Dynamic -> SomeDims -> IO ()

resize a dynamic tensor inplace with runtime SomeDims representation of its new shape.

getDim' :: Dynamic -> SomeDims -> IO HsReal

get a specific value of a dynamic tensor with runtime SomeDims index

new' :: SomeDims -> IO Dynamic

build a new tensor with a runtime SomeDims

resizeAs

Arguments

:: Dynamic
src

-> Dynamic
a tensor only used for its shape

-> IO Dynamic
a new copy of src with the shape tensor's shape

resize a tensor to take the shape of the second Dynamic argument. This is a pure function.

FIXME: Is this right? why are there three tensors

showTensor :: forall a. (Typeable a, Ord a, Num a, Show a) => (Int64 -> IO a) -> (Int64 -> Int64 -> IO a) -> (Int64 -> Int64 -> Int64 -> IO a) -> (Int64 -> Int64 -> Int64 -> Int64 -> IO a) -> [Int64] -> IO (String, String)

Generic way of showing the internal data of a tensor in a tabular format. This makes no assumptions about the type of representation to show and can be used for Storage, Dynamic, and Tensor types.

Helper functions

withDynamicStateAndStorage :: Dynamic -> Storage -> (Ptr CState -> Ptr CTensor -> Ptr CStorage -> IO x) -> IO x

run a function with a dynamic tensor and storage's underlying implementation details.

withInplace :: (Dynamic -> IO ()) -> Dims (d :: [Nat]) -> IO Dynamic

exported helper function. Not actually "inplace" this is actually "with return and static dimensions"

withInplace' :: (Dynamic -> IO ()) -> SomeDims -> IO Dynamic

exported helper function. not actually "inplace" this is actually "with return and runtime dimensions"

twice :: Dynamic -> (Dynamic -> Dynamic -> IO ()) -> IO Dynamic

exported helper function. This is actually inplace

withEmpty :: Dynamic -> (Dynamic -> IO ()) -> IO Dynamic

exported helper function. Should be renamed to newFromSize

withEmpty' :: (Dynamic -> IO ()) -> IO Dynamic

exported helper function. We can get away with this some of the time, when Torch does the resizing in C, but you need to look at the c implementation

Orphan instances

Show Dynamic

Instance details

Methods

showsPrec :: Int -> Dynamic -> ShowS

show :: Dynamic -> String

showList :: [Dynamic] -> ShowS