Torch.Indef.Dynamic.Tensor.Masked

Description

Operations using a mask tensor to filter which elements will be used.

Documentation

_maskedFill :: Dynamic -> MaskDynamic -> HsReal -> IO ()

fill a dynamic tensor with a value, filtered by a boolean mask tensor

C-Style: In the classic Torch C-style, the first argument is treated as the return type and is mutated in-place.

_maskedCopy :: Dynamic -> MaskDynamic -> Dynamic -> IO ()

copy a dynamic tensor with a value, filtered by a boolean mask tensor

C-Style: In the classic Torch C-style, the first argument is treated as the return type and is mutated in-place.

_maskedSelect :: Dynamic -> Dynamic -> MaskDynamic -> IO ()

select a dynamic tensor with a value, filtered by a boolean mask tensor

C-Style: In the classic Torch C-style, the first argument is treated as the return type and is mutated in-place.