Torch.Indef.Dynamic.Tensor.Math.Compare

Description

Compare a tensor with a scala value

Documentation

ltValue :: Dynamic -> HsReal -> MaskDynamic

return a byte tensor which contains boolean values indicating the relation between a tensor and a given scalar.

ltValueT :: Dynamic -> HsReal -> IO Dynamic

return a tensor which contains numeric values indicating the relation between a tensor and a given scalar. 0 stands for false, 1 stands for true.

ltValueT_ :: Dynamic -> HsReal -> IO ()

mutate a tensor in-place with its numeric relation to a given scalar, where 0 stands for false and 1 stands for true.

leValue :: Dynamic -> HsReal -> MaskDynamic

return a byte tensor which contains boolean values indicating the relation between a tensor and a given scalar.

leValueT :: Dynamic -> HsReal -> IO Dynamic

return a tensor which contains numeric values indicating the relation between a tensor and a given scalar. 0 stands for false, 1 stands for true.

leValueT_ :: Dynamic -> HsReal -> IO ()

mutate a tensor in-place with its numeric relation to a given scalar, where 0 stands for false and 1 stands for true.

gtValue :: Dynamic -> HsReal -> MaskDynamic

return a byte tensor which contains boolean values indicating the relation between a tensor and a given scalar.

gtValueT :: Dynamic -> HsReal -> IO Dynamic

return a tensor which contains numeric values indicating the relation between a tensor and a given scalar. 0 stands for false, 1 stands for true.

gtValueT_ :: Dynamic -> HsReal -> IO ()

mutate a tensor in-place with its numeric relation to a given scalar, where 0 stands for false and 1 stands for true.

geValue :: Dynamic -> HsReal -> MaskDynamic

return a byte tensor which contains boolean values indicating the relation between a tensor and a given scalar.

geValueT :: Dynamic -> HsReal -> IO Dynamic

return a tensor which contains numeric values indicating the relation between a tensor and a given scalar. 0 stands for false, 1 stands for true.

geValueT_ :: Dynamic -> HsReal -> IO ()

mutate a tensor in-place with its numeric relation to a given scalar, where 0 stands for false and 1 stands for true.

neValue :: Dynamic -> HsReal -> MaskDynamic

return a byte tensor which contains boolean values indicating the relation between a tensor and a given scalar.

neValueT :: Dynamic -> HsReal -> IO Dynamic

return a tensor which contains numeric values indicating the relation between a tensor and a given scalar. 0 stands for false, 1 stands for true.

neValueT_ :: Dynamic -> HsReal -> IO ()

mutate a tensor in-place with its numeric relation to a given scalar, where 0 stands for false and 1 stands for true.

eqValue :: Dynamic -> HsReal -> MaskDynamic

return a byte tensor which contains boolean values indicating the relation between a tensor and a given scalar.

eqValueT :: Dynamic -> HsReal -> IO Dynamic

return a tensor which contains numeric values indicating the relation between a tensor and a given scalar. 0 stands for false, 1 stands for true.

eqValueT_ :: Dynamic -> HsReal -> IO ()

mutate a tensor in-place with its numeric relation to a given scalar, where 0 stands for false and 1 stands for true.