Torch.Indef.Dynamic.Tensor.Math.Pairwise

Description

Documentation

equal :: Dynamic -> Dynamic -> IO Bool

Call Torch's C-level equal function.

add :: Dynamic -> HsReal -> IO Dynamic

add a scalar to a tensor.

add_ :: Dynamic -> HsReal -> IO ()

add a scalar to a tensor, inplace.

add_scaled_

Arguments

:: Dynamic
tensor to scale

-> HsReal
value to add

-> HsReal
amount to scale the value by

-> IO ()

add a scalar, which has been scaled, to a tensor, inplace.

sub :: Dynamic -> HsReal -> IO Dynamic

subtract a scalar from a tensor.

sub_ :: Dynamic -> HsReal -> IO ()

subtract a scalar from a tensor, inplace.

sub_scaled_

Arguments

:: Dynamic
tensor to scale

-> HsReal
value to add

-> HsReal
amount to scale the value by

-> IO ()

subtract a scalar, which has been scaled, from a tensor, inplace.

mul :: Dynamic -> HsReal -> IO Dynamic

multiply a tensor by a scalar value, pure.

mul_ :: Dynamic -> HsReal -> IO ()

multiply a tensor by a scalar value, inplace.

div :: Dynamic -> HsReal -> IO Dynamic

divide a tensor by a scalar value, pure.

div_ :: Dynamic -> HsReal -> IO ()

divide a tensor by a scalar value, inplace.

lshift_ :: Dynamic -> HsReal -> IO ()

Left shift all elements in the tensor by the given value, inplace.

rshift_ :: Dynamic -> HsReal -> IO ()

Right shift all elements in the tensor by the given value, inplace.

fmod_ :: Dynamic -> HsReal -> IO ()

Compute the remainder of division ( rounded towards zero) of all elements in the tensor by a given value, inplace.

remainder_ :: Dynamic -> HsReal -> IO ()

Computes remainder of division (rounded to nearest) of all elements in the tensor by value, inplace

bitand_ :: Dynamic -> HsReal -> IO ()

Performs the bitwise operation inplace on all elements in the tensor.

bitor_ :: Dynamic -> HsReal -> IO ()

Performs the bitwise operation inplace on all elements in the tensor.

bitxor_ :: Dynamic -> HsReal -> IO ()

Performs the bitwise operation inplace on all elements in the tensor.