Torch.Indef.Dynamic.Tensor.Math.Pointwise.Floating

Description

Documentation

cinv_ :: Dynamic -> IO ()

Inplace version of cinv, mutating the first tensor argument.

cinv :: Dynamic -> IO Dynamic

Return a new tensor applying 1.0 / x to all elements.

sigmoid_ :: Dynamic -> IO ()

Inplace version of sigmoid, mutating the first tensor argument.

sigmoid :: Dynamic -> IO Dynamic

Returns a new Tensor with the sigmoid of the elements of x.

log_ :: Dynamic -> IO ()

Inplace version of log, mutating the first tensor argument.

log :: Dynamic -> IO Dynamic

Returns a new tensor with the natural logarithm of the elements of x.

lgamma_ :: Dynamic -> IO ()

Inplace version of lgamma, mutating the first tensor argument.

lgamma :: Dynamic -> IO Dynamic

Returns a new tensor with the natural logarithm of the absolute value of the gamma function of the elements of x.

log1p_ :: Dynamic -> IO ()

Inplace version of log1p, mutating the first tensor argument.

log1p :: Dynamic -> IO Dynamic

Returns a new tensor with the natural logarithm of the elements of x + 1.

This function is more accurate than log for small values of x.

exp_ :: Dynamic -> IO ()

Inplace version of exp, mutating the first tensor argument.

exp :: Dynamic -> IO Dynamic

Returns, for each element in x, e (Neper number, the base of natural logarithms) raised to the power of the element in x.

cos_ :: Dynamic -> IO ()

Inplace version of cos, mutating the first tensor argument.

cos :: Dynamic -> IO Dynamic

Returns a new tensor with the cosine of the elements of x.

acos_ :: Dynamic -> IO ()

Inplace version of acos, mutating the first tensor argument.

acos :: Dynamic -> IO Dynamic

Returns a new tensor with the arcosine of the elements of x.

cosh_ :: Dynamic -> IO ()

Inplace version of cosh, mutating the first tensor argument.

cosh :: Dynamic -> IO Dynamic

Returns a new tensor with the hyberbolic cosine of the elements of x.

sin_ :: Dynamic -> IO ()

Inplace version of sin, mutating the first tensor argument.

sin :: Dynamic -> IO Dynamic

Returns a new tensor with the sine of the elements of x.

asin_ :: Dynamic -> IO ()

Inplace version of asin, mutating the first tensor argument.

asin :: Dynamic -> IO Dynamic

Returns a new tensor with the arcsine of the elements of x.

sinh_ :: Dynamic -> IO ()

Inplace version of sinh, mutating the first tensor argument.

sinh :: Dynamic -> IO Dynamic

Returns a new tensor with the hyperbolic sine of the elements of x.

tan_ :: Dynamic -> IO ()

Inplace version of tan, mutating the first tensor argument.

tan :: Dynamic -> IO Dynamic

Returns a new tensor with the tangent of the elements of x.

atan_ :: Dynamic -> IO ()

Inplace version of atan, mutating the first tensor argument.

atan :: Dynamic -> IO Dynamic

Returns a new tensor with the arctangent of the elements of x.

tanh_ :: Dynamic -> IO ()

Inplace version of tanh, mutating the first tensor argument.

tanh :: Dynamic -> IO Dynamic

Returns a new tensor with the hyperbolic tangent of the elements of x.

erf_ :: Dynamic -> IO ()

Inplace version of erf, mutating the first tensor argument.

erf :: Dynamic -> IO Dynamic

Returns a new tensor with the gauss error function applied to the elements of x.

The error function comes from https://en.wikipedia.org/wiki/Error_function

erfinv_ :: Dynamic -> IO ()

Inplace version of erfinv, mutating the first tensor argument.

erfinv :: Dynamic -> IO Dynamic

Returns a new tensor with the inverse gauss error function applied to the elements of x.

See https://en.wikipedia.org/wiki/Error_function for the gauss error function. This is its inverse.

pow_ :: Dynamic -> HsReal -> IO ()

Inplace version of pow, mutating the first tensor argument.

pow :: Dynamic -> HsReal -> IO Dynamic

Returns a new tensor with the elements of x to the power of n.

tpow_ :: HsReal -> Dynamic -> IO ()

Inplace version of tpow, mutating the first tensor argument.

tpow

Arguments

:: HsReal
base scalar n

-> Dynamic
tensor x of powers to raise n by.

-> IO Dynamic

Returns a new tensor with the scalar n, raised to the power of each element in the tensor x.

sqrt_ :: Dynamic -> IO ()

Inplace version of sqrt, mutating the first tensor argument.

sqrt :: Dynamic -> IO Dynamic

Returns a new tensor with the square root of the elements of x.

rsqrt_ :: Dynamic -> IO ()

Inplace version of rsqrt, mutating the first tensor argument.

rsqrt :: Dynamic -> IO Dynamic

Returns a new tensor with the reciprocal of the square root of the elements of x.

ceil_ :: Dynamic -> IO ()

Inplace version of ceil, mutating the first tensor argument.

ceil :: Dynamic -> IO Dynamic

Returns a new tensor with the values of the elements of x rounded up to the nearest integers.

floor_ :: Dynamic -> IO ()

Inplace version of floor, mutating the first tensor argument.

floor :: Dynamic -> IO Dynamic

Returns a new Tensor with the values of the elements of x rounded down to the nearest integers.

round_ :: Dynamic -> IO ()

Inplace version of round, mutating the first tensor argument.

round :: Dynamic -> IO Dynamic

Returns a new tensor with the values of the elements of x rounded to the nearest integers.

FIXME: The lua docs don't state how this rounding works. Someone should read the source code and document this.

trunc_ :: Dynamic -> IO ()

Inplace version of trunc, mutating the first tensor argument.

trunc :: Dynamic -> IO Dynamic

Returns a new tensor with the truncated integer values of the elements of x.

FIXME: The lua docs don't state how this truncation works. Someone should read the source code, document this, and explain how this differs from floor.

frac_ :: Dynamic -> IO ()

Inplace version of frac, mutating the first tensor argument.

frac :: Dynamic -> IO Dynamic

Returns a new tensor with the fractional portion of the elements of x.

lerp_ :: Dynamic -> Dynamic -> HsReal -> IO ()

Inplace version of lerp, mutating the first tensor argument.

lerp :: Dynamic -> Dynamic -> HsReal -> IO Dynamic

Linear interpolation of two scalars or tensors based on a weight:

  res = a + weight * (b - a)

atan2_ :: Dynamic -> Dynamic -> IO ()

Inplace version of atan2, mutating the first tensor argument.

atan2 :: Dynamic -> Dynamic -> IO Dynamic

Returns a new tensor with the arctangent of the elements of x and y. Note that the arctangent of the elements x and y refers to the signed angle in radians between the rays ending at origin where the first one starts at (1, 0) and the second at (y, x).