Torch.Indef.Dynamic.Tensor.Math.Reduce

Description

Documentation

minall :: Dynamic -> HsReal

get the minima of a tensor's elements

maxall :: Dynamic -> HsReal

get the maxima of a tensor's elements

medianall :: Dynamic -> HsReal

get the median value of a tensor's elements

sumall :: Dynamic -> HsAccReal

get the sum of a tensor's elements

prodall :: Dynamic -> HsAccReal

get the product of a tensor's elements

_max :: (Dynamic, IndexDynamic) -> Dynamic -> DimVal -> Maybe KeepDim -> IO ()

get the maximal value in the specified dimension and a corresponding index tensor of the maximum value's index.

Inplace and C-Style mutation

_min :: (Dynamic, IndexDynamic) -> Dynamic -> DimVal -> Maybe KeepDim -> IO ()

get the minimal value in the specified dimension and a corresponding index tensor of the minimum value's index.

Inplace and C-Style mutation

_median :: (Dynamic, IndexDynamic) -> Dynamic -> DimVal -> Maybe KeepDim -> IO ()

get the median value in the specified dimension and a corresponding index tensor of the median value's index.

Inplace and C-Style mutation

_sum :: Dynamic -> Dynamic -> DimVal -> Maybe KeepDim -> IO ()

sum the tensor in the specified dimension.

Inplace and C-Style mutation

_prod :: Dynamic -> Dynamic -> DimVal -> Maybe KeepDim -> IO ()

take the product of the tensor in the specified dimension.

Inplace and C-Style mutation

max :: Dynamic -> DimVal -> Maybe KeepDim -> IO (Dynamic, Maybe IndexDynamic)

get the maximum value in the specified dimension and return an optional corresponding index tensor of the maximum value's index.

min :: Dynamic -> DimVal -> Maybe KeepDim -> IO (Dynamic, Maybe IndexDynamic)

get the minimum value in the specified dimension and return an optional corresponding index tensor of the minimum value's index.

median :: Dynamic -> DimVal -> Maybe KeepDim -> IO (Dynamic, Maybe IndexDynamic)

get the median value in the specified dimension and return an optional corresponding index tensor of the median value's index.