Torch.Indef.Dynamic.Tensor.Math.Reduce.Floating

Description

Documentation

mean :: Dynamic -> DimVal -> IO Dynamic

Take the mean in the specified dimension.

mean_ :: Dynamic -> DimVal -> IO ()

Inplace mean

std :: Dynamic -> DimVal -> KeepDim -> Bool -> IO Dynamic

Performs the std operation over the specified dimension. The Bool parameter specifies whether the standard deviation should be used with n-1 or n. False normalizes by n-1, while True normalizes n.

std_ :: Dynamic -> DimVal -> KeepDim -> Bool -> IO ()

Infix version of std.

var :: Dynamic -> DimVal -> KeepDim -> Bool -> IO Dynamic

Get the variance over a tensor in the specified dimension. The Bool parameter specifies whether the standard deviation should be used with n-1 or n. False normalizes by n-1, while True normalizes n.

var_ :: Dynamic -> DimVal -> KeepDim -> Bool -> IO ()

Infix version of var.

norm :: Dynamic -> HsReal -> DimVal -> IO Dynamic

Return the p-norms of the tensor, computed over dimension dim.

norm_ :: Dynamic -> HsReal -> DimVal -> IO ()

Inplace version of norm

renorm

Arguments

:: Dynamic
x

-> HsReal
p

-> Int
dim

-> HsReal
maxnorm

-> IO Dynamic
res

Renormalizes the sub-Tensors along dimension dim such that they do not exceed norm maxnorm.

Equivalent to the following lua code: y = torch.renorm(x, p, dim, maxnorm). Returns a version of x with p-norms lower than maxnorm over non-dim dimensions. The dim argument is not to be confused with the argument of the same name in function norm. In this case, the p-norm is measured for each i-th sub-tensor (lua: x:select(dim, i)).

renorm_ :: Dynamic -> HsReal -> Int -> HsReal -> IO ()

inplace version of renorm

dist

Arguments

:: Dynamic
tensor x

-> Dynamic
tensor y

-> HsReal
p

-> IO HsAccReal

Returns the p-norm of x - y.

meanall :: Dynamic -> IO HsAccReal

Returns the mean of all elements.

varall :: Dynamic -> Int -> IO HsAccReal

Returns the variance of all elements.

stdall :: Dynamic -> Int -> IO HsAccReal

Returns the standard deviation of all elements.

normall

Arguments

:: Dynamic
tensor of values to norm over

-> HsReal
p

-> IO HsAccReal

Returns the p-norm of all elements.