Torch.Indef.Dynamic.Tensor.Sort

Description

Documentation

_sort :: (Dynamic, IndexDynamic) -> Dynamic -> DimVal -> SortOrder -> IO ()

Returns a tensor and index where all entries are sorted along the given dimension, in the chosen sort order. The index corresponds to the original indices.