Torch.Indef.Static.NN

Description

Documentation

_pReLU_updateOutput :: Tensor d -> Tensor d -> Tensor d -> IO ()

pReLU updateOutput

_pReLU_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

pReLU updateGradInput

_pReLU_accGradParameters :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> IO ()

pReLU accGradParameters

_rReLU_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Double -> Double -> Bool -> Bool -> Generator -> IO ()

rReLU updateOutput

_rReLU_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> Double -> Bool -> Bool -> IO ()

rReLU updateGradInput

_eLU_updateOutput :: Tensor d -> Tensor d -> Double -> Double -> Bool -> IO ()

eLU updateOutput

_eLU_updateGradInput :: Tensor d -> Tensor d' -> Tensor d'' -> Double -> Double -> IO ()

eLU updateGradInput

_leakyReLU_updateOutput :: Tensor d -> Tensor d -> Double -> Bool -> IO ()

leakyReLU updateOutput

_leakyReLU_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Double -> Bool -> IO ()

leakyReLU updateGradInput

relu :: Reifies s W => Dimensions d => BVar s (Tensor d) -> BVar s (Tensor d)

ReLU activation function

threshold

Arguments

:: Reifies s W

=> Dimensions d

=> Double
threshold

-> Double
replacement value

-> BVar s (Tensor d)
input

-> BVar s (Tensor d)
output

run a threshold function againts two BVar variables

_temporalUpSamplingNearest_updateOutput :: Tensor d -> Tensor d -> Int -> IO ()

temporalUpSamplingNearest forward pass (updates the output tensor)

_temporalUpSamplingNearest_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

temporalUpSamplingNearest backward-update (updates the layer and bias tensors)

_temporalUpSamplingLinear_updateOutput :: Tensor d -> Tensor d -> Int -> IO ()

temporalUpSamplingLinear forward pass (updates the output tensor)

_temporalUpSamplingLinear_updateGradInput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

temporalUpSamplingLinear backward-update (updates the layer and bias tensors)

_spatialSubSampling_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

spatialSubSampling forward pass (updates the output tensor)

_spatialSubSampling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

spatialSubSampling backward-update (updates the layer and bias tensors)

_spatialSubSampling_accGradParameters :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Double -> IO ()

spatialSubSampling backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_spatialUpSamplingNearest_updateOutput :: Tensor d -> Tensor d -> Int -> IO ()

spatialUpSamplingNearest forward pass (updates the output tensor)

_spatialUpSamplingNearest_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

spatialUpSamplingNearest backward-update (updates the layer and bias tensors)

_spatialUpSamplingBilinear_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> IO ()

spatialUpSamplingBilinear forward pass (updates the output tensor)

_spatialUpSamplingBilinear_updateGradInput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

spatialUpSamplingBilinear backward-update (updates the layer and bias tensors)

_spatialGridSamplerBilinear_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

spatialGridSamplerBilinear forward pass (updates the output tensor)

_spatialGridSamplerBilinear_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

spatialGridSamplerBilinear backward-update (updates the layer and bias tensors)

_volumetricGridSamplerBilinear_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

volumetricGridSamplerBilinear forward pass (updates the output tensor)

_volumetricGridSamplerBilinear_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

volumetricGridSamplerBilinear backward-update (updates the layer and bias tensors)

_volumetricUpSamplingNearest_updateOutput :: Tensor d -> Tensor d -> Int -> IO ()

volumetricUpSamplingNearest forward pass (updates the output tensor)

_volumetricUpSamplingNearest_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

volumetricUpSamplingNearest backward-update (updates the layer and bias tensors)

_volumetricUpSamplingTrilinear_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> IO ()

volumetricUpSamplingTrilinear forward pass (updates the output tensor)

_volumetricUpSamplingTrilinear_updateGradInput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricUpSamplingTrilinear backward-update (updates the layer and bias tensors)

_spatialReflectionPadding_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

spatialReflectionPadding forward pass (updates the output tensor)

_spatialReflectionPadding_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

spatialReflectionPadding backward-update (updates the layer and bias tensors)

_spatialReplicationPadding_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

spatialReplicationPadding forward pass (updates the output tensor)

_spatialReplicationPadding_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IO ()

spatialReplicationPadding backward-update (updates the layer and bias tensors)

_volumetricReplicationPadding_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricReplicationPadding forward pass (updates the output tensor)

_volumetricReplicationPadding_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricReplicationPadding backward-update (updates the layer and bias tensors)

_temporalReflectionPadding_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> IO ()

temporalReflectionPadding forward pass (updates the output tensor)

_temporalReflectionPadding_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> IO ()

temporalReflectionPadding backward-update (updates the layer and bias tensors)

_temporalReplicationPadding_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> IO ()

temporalReplicationPadding forward pass (updates the output tensor)

_temporalReplicationPadding_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> IO ()

temporalReplicationPadding backward-update (updates the layer and bias tensors)

abs_updateOutput :: Tensor d -> IO (Tensor d)

abs forward pass (updates the output tensor)

abs_updateGradInput

Arguments

:: Product d ~ Product d'

=> Tensor d
input

-> Tensor d'
gradOutput

-> IO (Tensor d)
gradInput

abs backward-update (updates the layer and bias tensors)

_sqrt_updateOutput :: Tensor d -> Tensor d -> Double -> IO ()

sqrt forward pass (updates the output tensor)

_sqrt_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

sqrt backward-update (updates the layer and bias tensors)

_square_updateOutput :: Tensor d -> Tensor d -> IO ()

square forward pass (updates the output tensor)

_square_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

square backward-update (updates the layer and bias tensors)

_logSigmoid_updateOutput :: Tensor d -> Tensor d -> Tensor d -> IO ()

logSigmoid forward pass (updates the output tensor)

_logSigmoid_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

logSigmoid backward-update (updates the layer and bias tensors)

_logSoftMax_updateOutput :: Tensor d -> Tensor d -> Integer -> IO ()

logSoftMax forward pass (updates the output tensor)

_logSoftMax_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Integer -> IO ()

logSoftMax backward-update (updates the layer and bias tensors)

_sigmoid_updateOutput :: Tensor d -> Tensor d -> IO ()

sigmoid forward pass (updates the output tensor)

_sigmoid_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

sigmoid backward-update (updates the layer and bias tensors)

_softMax_updateOutput :: Tensor d -> Tensor d -> Integer -> IO ()

softMax forward pass (updates the output tensor)

_softMax_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Integer -> IO ()

softMax backward-update (updates the layer and bias tensors)

_softPlus_updateOutput :: Tensor d -> Tensor d -> Double -> Double -> IO ()

softPlus forward pass (updates the output tensor)

_softPlus_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> Double -> IO ()

softPlus backward-update (updates the layer and bias tensors)

_softShrink_updateOutput :: Tensor d -> Tensor d -> Double -> IO ()

softShrink forward pass (updates the output tensor)

_softShrink_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Double -> IO ()

softShrink backward-update (updates the layer and bias tensors)

_tanh_updateOutput :: Tensor d -> Tensor d -> IO ()

tanh forward pass (updates the output tensor)

_tanh_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

tanh backward-update (updates the layer and bias tensors)

_hardTanh_updateOutput :: Tensor d -> Tensor d -> Double -> Double -> Bool -> IO ()

hardTanh forward pass (updates the output tensor)

_hardTanh_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Double -> Double -> Bool -> IO ()

hardTanh backward-update (updates the layer and bias tensors)

newtype Linear i o

datatype representing a linear layer with bias. Represents y = Ax + b.

Constructors

Linear (Tensor '[i, o], Tensor '[o])

linear :: forall s i o. Reifies s W => All KnownDim '[i, o] => BVar s (Linear i o) -> BVar s (Tensor '[i]) -> BVar s (Tensor '[o])

Backprop linear function without batching

flattenBP :: (Reifies s W, KnownDim (Product d), Dimensions (d :: [Nat])) => BVar s (Tensor d) -> BVar s (Tensor '[Product d])

A backpropable flatten operation

_sparseLinear_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

sparseLinear forward pass (updates the output tensor)

_sparseLinear_accGradParameters :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> Double -> IO ()

sparseLinear backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_sparseLinear_zeroGradParameters :: Tensor d -> Tensor d -> Tensor d -> IO ()

sparseLinear zeroGradParameters

_sparseLinear_updateParameters :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> IO ()

sparseLinear updateParameters

_gatedLinear_updateOutput :: Tensor d -> Tensor d -> Int -> IO ()

gatedLinear forward pass (updates the output tensor)

_gatedLinear_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> IO ()

gatedLinear backward-update (updates the layer and bias tensors)

_gRUFused_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

gRUFused forward pass (updates the output tensor)

_gRUFused_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

gRUFused backward-update (updates the layer and bias tensors)

_lSTMFused_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

lSTMFused forward pass (updates the output tensor)

_lSTMFused_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> Tensor d -> IO ()

lSTMFused backward-update (updates the layer and bias tensors)

_absCriterion_updateOutput

Arguments

:: Tensor d
input

-> Tensor d'
target

-> Tensor d''
output

-> Bool
size average

-> Bool
reduce

-> IO ()

absCriterion forward pass (updates the output tensor)

_absCriterion_updateGradInput

Arguments

:: Tensor d
input

-> Tensor d'
target

-> Tensor d''
gradOutput

-> Tensor d''
gradInput

-> Bool
size average

-> Bool
reduce

-> IO ()

absCriterion backward-update (updates the layer and bias tensors)

_bCECriterion_updateOutput

Arguments

:: Tensor d
input

-> Tensor d'
target

-> Tensor d''
output

-> Bool
sizeAverage

-> Tensor d'''
weights

-> Bool
reduce

-> IO ()

bCECriterion forward pass (updates the output tensor)

_bCECriterion_updateGradInput

Arguments

:: Tensor d
input

-> Tensor d'
target

-> Tensor d''
grad output

-> Tensor d'''
grad input

-> Bool
sizeAvreage

-> Tensor d''''
weights

-> Bool
reduce

-> IO ()

bCECriterion backward-update (updates the layer and bias tensors)

_marginCriterion_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Bool -> Double -> IO ()

marginCriterion forward pass (updates the output tensor)

_marginCriterion_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Bool -> Double -> IO ()

marginCriterion backward-update (updates the layer and bias tensors)

_softMarginCriterion_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

softMarginCriterion forward pass (updates the output tensor)

_softMarginCriterion_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

softMarginCriterion backward-update (updates the layer and bias tensors)

_mSECriterion_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

mSECriterion forward pass (updates the output tensor)

_mSECriterion_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

mSECriterion backward-update (updates the layer and bias tensors)

_distKLDivCriterion_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

distKLDivCriterion forward pass (updates the output tensor)

_distKLDivCriterion_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

distKLDivCriterion backward-update (updates the layer and bias tensors)

_smoothL1Criterion_updateOutput :: Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

smoothL1Criterion forward pass (updates the output tensor)

_smoothL1Criterion_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Bool -> Bool -> IO ()

smoothL1Criterion backward-update (updates the layer and bias tensors)

_l1Cost_updateOutput :: Tensor d -> Tensor d -> IO ()

l1Cost forward pass (updates the output tensor)

_l1Cost_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

l1Cost backward-update (updates the layer and bias tensors)

type SideCheck h k d p o = (All KnownDim '[h, k, d, p, o], (k > 0) ~ True, (d > 0) ~ True, ((h + (2 * p)) < k) ~ False, (o > 0) ~ True, o ~ (Div ((h + (2 * p)) - k) d + 1))

Constraint to check valid dimensions on one side.

type SpatialConvolutionC f h w kH kW dH dW pH pW oH oW = (All KnownDim '[(f * kH) * kW, oH * oW, f], SideCheck h kH dH pH oH, SideCheck w kW dW pW oW)

Constraint to check both sides (height and width) of a function and assert that all nessecary dimension values are KnownDims.

data Dilation2d (h :: Nat) (w :: Nat)

Representation of how much to dilate in the height and width dimensions

Constructors

Dilation2d

data Kernel2d (h :: Nat) (w :: Nat)

Representation of how big a kernel will be in the height and width dimensions

Constructors

Kernel2d

data Padding2d (h :: Nat) (w :: Nat)

Representation of how much to pad in the height and width dimensions

Constructors

Padding2d

data Step2d (h :: Nat) (w :: Nat)

Representation of how much to step in the height and width dimensions

Constructors

Step2d

class Param2d (p :: Nat -> Nat -> Type) where

Typeclass to generically pull out Width and Height information from a parameter

FIXME: this can be replaced with simple functions.

Methods

paramW :: forall w h i. (KnownDim w, Integral i) => p h w -> i

get the width parameter

paramH :: forall w h i. (KnownDim h, Integral i) => p h w -> i

get the height parameter

param2d :: (KnownDim h, KnownDim w, Integral i) => p h w -> (i, i)

get both parameters as a (width, height) tuple FIXME: Isn't this supposed to be "height" then "width"???

newtype Conv2d i o kH kW

ADT representation of a convolutional 2d layer.

FIXME: the type is a bit of a hiccup: can we remove the kernel dimensions or move pad/stride into the phantoms?

possibly something like Conv2d i o (kH, kW) (dH, dW) (pH, pW) or Conv2d i o (kH, kW) (Maybe (dH, dW)) (Maybe (pH, pW))

Constructors

Conv2d (Tensor '[o, i, kH, kW], Tensor '[o])

kernel2d :: (Integral i, KnownDim kH, KnownDim kW) => Conv2d f o kH kW -> (i, i)

get the kernel tuple as (width, height) from a Conv2d ADT

FIXME: Isn't this supposed to be "height" then "width"???

conv2dMM

Arguments

:: Reifies s W

=> SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[f, o]

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Double
learning rate

-> BVar s (Conv2d f o kH kW)
conv2d state

-> BVar s (Tensor '[f, h, w])
input: f stands for "features" or "input plane")

-> BVar s (Tensor '[o, oH, oW])

Backprop convolution function

conv2dMMBatch

Arguments

:: Reifies s W

=> SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[f, o, b]

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Double
learning rate

-> BVar s (Conv2d f o kH kW)
conv2d state

-> BVar s (Tensor '[b, f, h, w])
input: f stands for "features" or "input plane")

-> BVar s (Tensor '[b, o, oH, oW])

Backprop convolution function with batching

_conv2dMM

Arguments

:: Reifies s W

=> All Dimensions '[din, dout, fgin]

=> All KnownDim '[f, o, kH, kW, dH, dW, pH, pW]

=> IO (Tensor fgin)
make grad input buffer

-> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Double
learning rate

-> BVar s (Conv2d f o kH kW)
conv2d state

-> BVar s (Tensor din)
input: f stands for "features" or "input plane")

-> BVar s (Tensor dout)

Backprop convolution function

conv2dMM_forward

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> KnownDim o

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[f, h, w]
input: f stands for "features" or "input plane"

-> Tensor '[o, oH, oW]

Applies a 2D convolution over an input image composed of several input planes. The input tensor in forward(input) is expected to be a 3D tensor (nInputPlane x height x width).

conv2dMM_updGradInput

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> KnownDim o

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[f, h, w]
input

-> Tensor '[o, oH, oW]
gradOutput

-> Tensor '[f, h, w]

conv2dMM updGradInput

conv2dMM_updGradParameters

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> KnownDim o

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW

-> Tensor '[f, h, w]
input

-> Tensor '[o, oH, oW]
gradOutput

-> Conv2d f o kH kW

conv2dMM updGradParameters

conv2dMM_forwardBatch

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[b, o]

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[b, f, h, w]
input: f stands for "features" or "input plane"

-> Tensor '[b, o, oH, oW]

conv2dMM_forward with a batch dimension

conv2dMM_updGradInputBatch

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[b, o, f * kW]

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[b, f, h, w]
input

-> Tensor '[b, o, oH, oW]
gradOutput

-> Tensor '[b, f, h, w]

conv2dMM_updGradInputBatch with batch dimension

conv2dMM_updGradParametersBatch

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[b, o]

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW

-> Tensor '[b, f, h, w]
input

-> Tensor '[b, o, oH, oW]
gradOutput

-> Conv2d f o kH kW

conv2dMM_updGradParameters with batch dimension

_conv2dMM_forward :: All KnownDim '[kH, kW, dH, dW, pH, pW, f, o] => Step2d dH dW -> Padding2d pH pW -> Conv2d f o kH kW -> Tensor din -> Tensor dout

helper of forward functions with unspecified dimensions

_conv2dMM_updGradInput :: forall f o oH oW kH kW dH dW pH pW inp gout fgin. All KnownDim '[kH, kW, dH, dW, pH, pW, f, o] => IO (Tensor fgin) -> Step2d dH dW -> Padding2d pH pW -> Conv2d f o kH kW -> Tensor inp -> Tensor gout -> Tensor inp

helper of backward update to compute gradient input with unspecified dimensions

_conv2dMM_accGradParameters

Arguments

:: All KnownDim '[kH, kW, dH, dW, pH, pW, f, o]

=> Dimensions finput

=> IO (Tensor finput)

-> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW
weights and bias which will be mutated in-place

-> Tensor inp
input

-> Tensor gout
gradOutput

-> IO ()

conv2dMM accGradParameters

_conv2dMM_updGradParameters

Arguments

:: All KnownDim '[kH, kW, dH, dW, pH, pW, f, o]

=> Dimensions finput

=> IO (Tensor finput)

-> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW
weights and bias which will be mutated in-place

-> Tensor inp
input

-> Tensor gout
gradOutput

-> Conv2d f o kH kW

conv2dMM updGradParameters

type SpatialDilationC iH iW kH kW dH dW pH pW oW oH dilH dilW ceilMode = (SpatialDilationCheckC kH kW dH dW pH pW dilH dilW, CeilModeOutputDims iH kH dH pH oH dilH ceilMode, CeilModeOutputDims iW kW dW pW oW dilW ceilMode, All KnownDim '[oH, oW, iH, iW])

Top-level constraint to assert that checks CeilModeOutputDims on height and width dimensions and asserts that all dimensions checks in SpatialDilationCheckC are true.

type CeilModeOutputDims i k d p o dil ceilMode = If (ceilMode && (Rem ((i - ((dil * (k - 1)) + 1)) + (2 * p)) d > 0)) ((2 + Div ((i - ((dil * (k - 1)) + 1)) + (2 * p)) d) ~ o) ((1 + Div ((i - ((dil * (k - 1)) + 1)) + (2 * p)) d) ~ o)

Type-level if statement to indicate what the output dimension should be if CeilMode is turned on.

type SpatialDilationCheckC kH kW dH dW pH pW dilH dilW = (All KnownDim '[kH, kW, pH, pW, dH, dW, dilH, dilW], (kW > 0) ~ True, (kH > 0) ~ True, (dW > 0) ~ True, (dH > 0) ~ True, (dilW > 0) ~ True, (dilH > 0) ~ True, (Div kW 2 >= pW) ~ True, (Div kH 2 >= pH) ~ True)

Constraint to assert that all hyperparameters are valid and to make the requirement that all dimension values are KnownDims.

_featureLPPooling_updateOutput :: Tensor d -> Tensor d -> Double -> Int -> Int -> Bool -> IO ()

featureLPPooling forward pass (updates the output tensor)

_featureLPPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> Int -> Int -> Bool -> IO ()

featureLPPooling backward-update (updates the layer and bias tensors)

_temporalMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

temporalMaxPooling forward pass (updates the output tensor)

_temporalMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

temporalMaxPooling backward-update (updates the layer and bias tensors)

dilatedMaxPooling2d

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH dilH dilW ceilMode

=> KnownDim inPlane

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> Dilation2d dilH dilW
dilation size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[inPlane, iW, iH])

-> BVar s (Tensor '[inPlane, oW, oH])

run a backprop-aware dilatedMaxPooling2d function

dilatedMaxPooling2dBatch

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH dilH dilW ceilMode

=> KnownDim inPlane

=> KnownDim b

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> Dilation2d dilH dilW
dilation size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[b, inPlane, iW, iH])

-> BVar s (Tensor '[b, inPlane, oW, oH])

run a backprop-aware dilatedMaxPooling2d function with a batch dimension.

_dilatedMaxPooling2d

Arguments

:: All KnownDim '[kH, kW, pH, pW, dH, dW, dilH, dilW]

=> All Dimensions '[d', d]

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> Dilation2d dilH dilW
dilation size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor d)
input

-> BVar s (Tensor d')
output

internal function of dilatedMaxPooling2d and dilatedMaxPooling2dBatch. Should not be used.

_maxPooling2d

Arguments

:: All KnownDim '[kH, kW, pH, pW, dH, dW]

=> All Dimensions '[d', d]

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size. Note: default in C is the kernel size.

-> Padding2d pH pW
padding size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor d)
input

-> BVar s (Tensor d')
output

internal function of maxPooling2d and maxPooling2dBatch. Should not be used.

maxPooling2d

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH 1 1 ceilMode

=> Reifies s W

=> KnownDim inPlane

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[inPlane, iH, iW])

-> BVar s (Tensor '[inPlane, oH, oW])

backprop-aware maxPooling2d function.

maxPooling2dBatch

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH 1 1 ceilMode

=> Reifies s W

=> KnownDim inPlane

=> KnownDim b

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[b, inPlane, iH, iW])

-> BVar s (Tensor '[b, inPlane, oH, oW])

backprop-aware maxPooling2d function with a batch dimension.

_spatialAdaptiveMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

spatialAdaptiveMaxPooling forward pass (updates the output tensor)

_spatialAdaptiveMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> IO ()

spatialAdaptiveMaxPooling backward-update (updates the layer and bias tensors)

_spatialFractionalMaxPooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IndexTensor d -> Tensor d -> IO ()

spatialFractionalMaxPooling forward pass (updates the output tensor)

_spatialFractionalMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IndexTensor d -> IO ()

spatialFractionalMaxPooling backward-update (updates the layer and bias tensors)

_spatialMaxUnpooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

spatialMaxUnpooling forward pass (updates the output tensor)

_spatialMaxUnpooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

spatialMaxUnpooling backward-update (updates the layer and bias tensors)

_spatialAdaptiveAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> IO ()

spatialAdaptiveAveragePooling forward pass (updates the output tensor)

_spatialAdaptiveAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

spatialAdaptiveAveragePooling backward-update (updates the layer and bias tensors)

_spatialAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

spatialAveragePooling forward pass (updates the output tensor)

_spatialAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

spatialAveragePooling backward-update (updates the layer and bias tensors)

_volumetricFractionalMaxPooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IndexTensor d -> Tensor d -> IO ()

volumetricFractionalMaxPooling forward pass (updates the output tensor)

_volumetricFractionalMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IndexTensor d -> IO ()

volumetricFractionalMaxPooling backward-update (updates the layer and bias tensors)

_volumetricMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricMaxPooling forward pass (updates the output tensor)

_volumetricMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricMaxPooling backward-update (updates the layer and bias tensors)

_volumetricDilatedMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricDilatedMaxPooling forward pass (updates the output tensor)

_volumetricDilatedMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricDilatedMaxPooling backward-update (updates the layer and bias tensors)

_volumetricMaxUnpooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricMaxUnpooling forward pass (updates the output tensor)

_volumetricMaxUnpooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricMaxUnpooling backward-update (updates the layer and bias tensors)

_volumetricAdaptiveMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> IO ()

volumetricAdaptiveMaxPooling forward pass (updates the output tensor)

_volumetricAdaptiveMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> IO ()

volumetricAdaptiveMaxPooling backward-update (updates the layer and bias tensors)

_volumetricAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

volumetricAveragePooling forward pass (updates the output tensor)

_volumetricAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

volumetricAveragePooling backward-update (updates the layer and bias tensors)

_volumetricAdaptiveAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> IO ()

volumetricAdaptiveAveragePooling forward pass (updates the output tensor)

_volumetricAdaptiveAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

volumetricAdaptiveAveragePooling backward-update (updates the layer and bias tensors)

newtype Conv1d f o kW dW

ADT representation of a convolutional 1d layer.

FIXME: the type is a bit of a hiccup: can we remove the kernel dimensions or move pad/stride into the phantoms?

See Conv2d for ideas.

Constructors

Conv1d (Tensor '[o, f * kW], Tensor '[o])

stepSize :: forall f o kW dW. KnownDim dW => Conv1d f o kW dW -> Int

dW: The step of the convolution. Default is 1 in C.

conv1d :: forall s seq f kW dW o. Reifies s W => KnownDim (f * kW) => TemporalConvC seq f kW dW o => Double -> BVar s (Conv1d f o kW dW) -> BVar s (Tensor '[seq, f]) -> BVar s (Tensor '[seq, o])

Backprop convolution function

conv1dBatch :: forall s seq f kW dW o b. Reifies s W => KnownDim b => KnownDim (f * kW) => TemporalConvC seq f kW dW o => Double -> BVar s (Conv1d f o kW dW) -> BVar s (Tensor '[b, seq, f]) -> BVar s (Tensor '[b, seq, o])

Backprop convolution function with batching

conv1d_forward :: TemporalConvC s f kW dW o => Conv1d f o kW dW -> Tensor '[s, f] -> IO (Tensor '[s, o])

If the input sequence is a 2D tensor of dimension (nInputFrame x inputFrameSize), the output sequence will be (nOutputFrame x outputFrameSize) where

nOutputFrame = (nInputFrame - kW) / dW + 1

conv1d_backwardGradInput

Arguments

:: TemporalConvC seq f kW dW o

=> Conv1d f o kW dW
conv1d state

-> Tensor '[seq, f]
input: s for 'sequence dimension', f for 'feature dimension'

-> Tensor '[seq, o]
grad output

-> IO (Tensor '[seq, f])
grad input

backward pass, computing the gradient input

conv1d_updGradParams

Arguments

:: TemporalConvC s f kW dW o

=> Conv1d f o kW dW
input state of conv1d (which includes weights and bias)

-> Tensor '[s, f]
input tensor

-> Tensor '[s, o]
output gradient

-> Double
scale

-> IO (Conv1d f o kW dW)
gradient of (weights, bias)

backward pass, computing the weight and bias parameters

WARNING: this is _pure_ which may be slow for large tensors. Speeding this up will be in active development as the need arises (see issue hasktorch/hasktorch#85)

conv1d_forwardBatch :: TemporalConvC s f kW dW o => Conv1d f o kW dW -> Tensor '[b, s, f] -> IO (Tensor '[b, s, o])

Applies a 1D convolution over an input sequence composed of nInputFrame frames. The input tensor in forward(input) is expected to be a 2D tensor (nInputFrame x inputFrameSize) or a 3D tensor (nBatchFrame x nInputFrame x inputFrameSize).

conv1d_backwardGradInputBatch

Arguments

:: TemporalConvC s f kW dW o

=> KnownDim b

=> Conv1d f o kW dW
conv1d state

-> Tensor '[b, s, f]
input: s for 'sequence dimension', f for 'feature dimension'

-> Tensor '[b, s, o]
grad output

-> IO (Tensor '[b, s, f])
output

conv1d_backwardGradInput with a batch dimension

conv1d_updGradParamsBatch

Arguments

:: TemporalConvC s f kW dW o

=> KnownDim b

=> Conv1d f o kW dW
conv1d state

-> Tensor '[b, s, f]
input: s for 'sequence dimension', f for 'feature dimension'

-> Tensor '[b, s, o]
grad output

-> Double
scale

-> IO (Conv1d f o kW dW)
output

conv1d_updGradParams with a batch dimension

_temporalRowConvolution_updateOutput :: Tensor d -> Tensor d' -> Tensor d'' -> Tensor d''' -> Tensor d -> Tensor d -> Int -> Int -> Int -> Bool -> IO ()

temporalRowConvolution forward pass (updates the output tensor)

_temporalRowConvolution_updateGradInput :: Tensor d -> Tensor d' -> Tensor d'' -> Tensor d''' -> Tensor d -> Tensor d -> Int -> Int -> Int -> Bool -> IO ()

temporalRowConvolution backward-update (updates the layer and bias tensors)

_temporalRowConvolution_updGradParameters :: Tensor d -> Tensor d' -> Tensor d'' -> Tensor d''' -> Tensor d -> Tensor d -> Int -> Int -> Int -> Bool -> Double -> IO ()

temporalRowConvolution backward-update (updates the layer and bias tensors). Called accGradParameters in C to indicate accumulating the gradient parameters.

_batchNormalization_updateOutput

Arguments

:: Tensor d
input

-> Tensor d
output

-> Tensor d
weight

-> Tensor d
bias

-> Tensor d
running mean

-> Tensor d
running var

-> Tensor d
save mean

-> Tensor d
save std

-> Bool
train

-> Double
momentum

-> Double
eps

-> IO ()

batchNormalization forward pass (updates the output tensor)

_batchNormalization_backward

Arguments

:: Tensor d
input

-> Tensor d
grad output

-> Tensor d
grad input

-> Tensor d
grad weight

-> Tensor d
grad bias

-> Tensor d
weight

-> Tensor d
running mean

-> Tensor d
running var

-> Tensor d
save mean

-> Tensor d
save std

-> Bool
train

-> Double
momentum

-> Double
eps

-> IO ()

batchNormalization backward

_col2Im_updateOutput

Arguments

:: Tensor d
input

-> Tensor d
output

-> Int
output Height

-> Int
output Width

-> Int
kH

-> Int
kW

-> Int
dH

-> Int
dW

-> Int
padH

-> Int
padW

-> Int
sH

-> Int
sW

-> IO ()

col2Im forward pass (updates the output tensor)

_col2Im_updateGradInput

Arguments

:: Tensor d
grad output

-> Tensor d
grad input

-> Int
kH

-> Int
kW

-> Int
dH

-> Int
dW

-> Int
padH

-> Int
padW

-> Int
sH

-> Int
sW

-> IO ()

col2Im backward-update (updates the layer and bias tensors)

_im2Col_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

im2Col forward pass (updates the output tensor)

_im2Col_updateGradInput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

im2Col backward-update (updates the layer and bias tensors)