Torch.Indef.Static.NN.Backprop

Description

Backprop helper instances for static tensors, as well as any helper functions that might work well with backprop.

Orphan instances

Dimensions d => Backprop (Tensor d)

Instance details

Methods

zero :: Tensor d -> Tensor d

add :: Tensor d -> Tensor d -> Tensor d

one :: Tensor d -> Tensor d