Torch.Indef.Static.NN.Conv2d

Description

Spatial (2D) Convolutions.

Excluding an optional batch dimension, spatial layers expect a 3D Tensor as input. The first dimension is the number of features (e.g. frameSize), the last two dimensions are spatial (e.g. height x width). These are commonly used for processing images.

Complete types and documentation at https://github.com/torch/nn/blob/master/doc/convolution.md#spatial-modules

Documentation

newtype Conv2d i o kH kW

ADT representation of a convolutional 2d layer.

FIXME: the type is a bit of a hiccup: can we remove the kernel dimensions or move pad/stride into the phantoms?

possibly something like Conv2d i o (kH, kW) (dH, dW) (pH, pW) or Conv2d i o (kH, kW) (Maybe (dH, dW)) (Maybe (pH, pW))

Constructors

Conv2d

Fields

getTensors :: (Tensor '[o, i, kH, kW], Tensor '[o])

weights :: Conv2d i o kH kW -> Tensor '[o, i, kH, kW]

get the weights from a Conv2d ADT

bias :: Conv2d i o kH kW -> Tensor '[o]

get the bias from a Conv2d ADT

featureSize :: forall i o kH kW. KnownDim i => Conv2d i o kH kW -> Int

get the featureSize from a Conv2d ADT

outputSize :: forall f o kH kW. KnownDim o => Conv2d f o kH kW -> Int

get the outputSize from a Conv2d ADT

kernelWidth :: forall i f o kH kW. (Integral i, KnownDim kW) => Conv2d f o kH kW -> i

get the kernelWidth from a Conv2d ADT

kernelHeight :: forall i f o kH kW. (Integral i, KnownDim kH) => Conv2d f o kH kW -> i

get the kernelHeight from a Conv2d ADT

kernel2d :: (Integral i, KnownDim kH, KnownDim kW) => Conv2d f o kH kW -> (i, i)

get the kernel tuple as (width, height) from a Conv2d ADT

FIXME: Isn't this supposed to be "height" then "width"???

class Param2d (p :: Nat -> Nat -> Type) where

Typeclass to generically pull out Width and Height information from a parameter

FIXME: this can be replaced with simple functions.

Methods

paramW :: forall w h i. (KnownDim w, Integral i) => p h w -> i

get the width parameter

paramH :: forall w h i. (KnownDim h, Integral i) => p h w -> i

get the height parameter

param2d :: (KnownDim h, KnownDim w, Integral i) => p h w -> (i, i)

get both parameters as a (width, height) tuple FIXME: Isn't this supposed to be "height" then "width"???

data Step2d (h :: Nat) (w :: Nat)

Representation of how much to step in the height and width dimensions

Constructors

Step2d

data Padding2d (h :: Nat) (w :: Nat)

Representation of how much to pad in the height and width dimensions

Constructors

Padding2d

data Kernel2d (h :: Nat) (w :: Nat)

Representation of how big a kernel will be in the height and width dimensions

Constructors

Kernel2d

data Dilation2d (h :: Nat) (w :: Nat)

Representation of how much to dilate in the height and width dimensions

Constructors

Dilation2d

type SpatialConvolutionC f h w kH kW dH dW pH pW oH oW = (All KnownDim '[(f * kH) * kW, oH * oW, f], SideCheck h kH dH pH oH, SideCheck w kW dW pW oW)

Constraint to check both sides (height and width) of a function and assert that all nessecary dimension values are KnownDims.

type SideCheck h k d p o = (All KnownDim '[h, k, d, p, o], (k > 0) ~ True, (d > 0) ~ True, ((h + (2 * p)) < k) ~ False, (o > 0) ~ True, o ~ (Div ((h + (2 * p)) - k) d + 1))

Constraint to check valid dimensions on one side.

conv2dMM

Arguments

:: Reifies s W

=> SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[f, o]

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Double
learning rate

-> BVar s (Conv2d f o kH kW)
conv2d state

-> BVar s (Tensor '[f, h, w])
input: f stands for "features" or "input plane")

-> BVar s (Tensor '[o, oH, oW])

Backprop convolution function

conv2dMMBatch

Arguments

:: Reifies s W

=> SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[f, o, b]

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Double
learning rate

-> BVar s (Conv2d f o kH kW)
conv2d state

-> BVar s (Tensor '[b, f, h, w])
input: f stands for "features" or "input plane")

-> BVar s (Tensor '[b, o, oH, oW])

Backprop convolution function with batching

_conv2dMM

Arguments

:: Reifies s W

=> All Dimensions '[din, dout, fgin]

=> All KnownDim '[f, o, kH, kW, dH, dW, pH, pW]

=> IO (Tensor fgin)
make grad input buffer

-> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Double
learning rate

-> BVar s (Conv2d f o kH kW)
conv2d state

-> BVar s (Tensor din)
input: f stands for "features" or "input plane")

-> BVar s (Tensor dout)

Backprop convolution function

conv2dMM_forward

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> KnownDim o

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[f, h, w]
input: f stands for "features" or "input plane"

-> Tensor '[o, oH, oW]

Applies a 2D convolution over an input image composed of several input planes. The input tensor in forward(input) is expected to be a 3D tensor (nInputPlane x height x width).

conv2dMM_updGradInput

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> KnownDim o

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[f, h, w]
input

-> Tensor '[o, oH, oW]
gradOutput

-> Tensor '[f, h, w]

conv2dMM updGradInput

conv2dMM_updGradParameters

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> KnownDim o

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW

-> Tensor '[f, h, w]
input

-> Tensor '[o, oH, oW]
gradOutput

-> Conv2d f o kH kW

conv2dMM updGradParameters

conv2dMM_forwardBatch

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[b, o]

=> Step2d dH dW
step of the convolution in width and height dimensions. C-default is 1 for both.

-> Padding2d pH pW
zero padding to the input plane for width and height. (kW-1)/2 is often used. C-default is 0 for both.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[b, f, h, w]
input: f stands for "features" or "input plane"

-> Tensor '[b, o, oH, oW]

conv2dMM_forward with a batch dimension

conv2dMM_updGradInputBatch

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[b, o, f * kW]

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height.

-> Conv2d f o kH kW
conv2d state

-> Tensor '[b, f, h, w]
input

-> Tensor '[b, o, oH, oW]
gradOutput

-> Tensor '[b, f, h, w]

conv2dMM_updGradInputBatch with batch dimension

conv2dMM_updGradParametersBatch

Arguments

:: SpatialConvolutionC f h w kH kW dH dW pH pW oH oW

=> All KnownDim '[b, o]

=> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW

-> Tensor '[b, f, h, w]
input

-> Tensor '[b, o, oH, oW]
gradOutput

-> Conv2d f o kH kW

conv2dMM_updGradParameters with batch dimension

_conv2dMM_forward :: All KnownDim '[kH, kW, dH, dW, pH, pW, f, o] => Step2d dH dW -> Padding2d pH pW -> Conv2d f o kH kW -> Tensor din -> Tensor dout

helper of forward functions with unspecified dimensions

_conv2dMM_updGradInput :: forall f o oH oW kH kW dH dW pH pW inp gout fgin. All KnownDim '[kH, kW, dH, dW, pH, pW, f, o] => IO (Tensor fgin) -> Step2d dH dW -> Padding2d pH pW -> Conv2d f o kH kW -> Tensor inp -> Tensor gout -> Tensor inp

helper of backward update to compute gradient input with unspecified dimensions

_conv2dMM_accGradParameters

Arguments

:: All KnownDim '[kH, kW, dH, dW, pH, pW, f, o]

=> Dimensions finput

=> IO (Tensor finput)

-> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW
weights and bias which will be mutated in-place

-> Tensor inp
input

-> Tensor gout
gradOutput

-> IO ()

conv2dMM accGradParameters

_conv2dMM_updGradParameters

Arguments

:: All KnownDim '[kH, kW, dH, dW, pH, pW, f, o]

=> Dimensions finput

=> IO (Tensor finput)

-> Step2d dH dW
(dH, dW) step of the convolution in width and height dimensions

-> Padding2d pH pW
(pH, pW) zero padding to the input plane for width and height. (kW-1)/2 is often used.

-> Double
scale / learning rate

-> Conv2d f o kH kW
weights and bias which will be mutated in-place

-> Tensor inp
input

-> Tensor gout
gradOutput

-> Conv2d f o kH kW

conv2dMM updGradParameters