Torch.Indef.Static.NN.Pooling

Description

Documentation

_featureLPPooling_updateOutput :: Tensor d -> Tensor d -> Double -> Int -> Int -> Bool -> IO ()

featureLPPooling forward pass (updates the output tensor)

_featureLPPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Tensor d -> Double -> Int -> Int -> Bool -> IO ()

featureLPPooling backward-update (updates the layer and bias tensors)

1d pooling functions

_temporalMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

temporalMaxPooling forward pass (updates the output tensor)

_temporalMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

temporalMaxPooling backward-update (updates the layer and bias tensors)

2d pooling functions

type SpatialDilationCheckC kH kW dH dW pH pW dilH dilW = (All KnownDim '[kH, kW, pH, pW, dH, dW, dilH, dilW], (kW > 0) ~ True, (kH > 0) ~ True, (dW > 0) ~ True, (dH > 0) ~ True, (dilW > 0) ~ True, (dilH > 0) ~ True, (Div kW 2 >= pW) ~ True, (Div kH 2 >= pH) ~ True)

Constraint to assert that all hyperparameters are valid and to make the requirement that all dimension values are KnownDims.

type CeilModeOutputDims i k d p o dil ceilMode = If (ceilMode && (Rem ((i - ((dil * (k - 1)) + 1)) + (2 * p)) d > 0)) ((2 + Div ((i - ((dil * (k - 1)) + 1)) + (2 * p)) d) ~ o) ((1 + Div ((i - ((dil * (k - 1)) + 1)) + (2 * p)) d) ~ o)

Type-level if statement to indicate what the output dimension should be if CeilMode is turned on.

type SpatialDilationC iH iW kH kW dH dW pH pW oW oH dilH dilW ceilMode = (SpatialDilationCheckC kH kW dH dW pH pW dilH dilW, CeilModeOutputDims iH kH dH pH oH dilH ceilMode, CeilModeOutputDims iW kW dW pW oW dilW ceilMode, All KnownDim '[oH, oW, iH, iW])

Top-level constraint to assert that checks CeilModeOutputDims on height and width dimensions and asserts that all dimensions checks in SpatialDilationCheckC are true.

dilatedMaxPooling2d

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH dilH dilW ceilMode

=> KnownDim inPlane

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> Dilation2d dilH dilW
dilation size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[inPlane, iW, iH])

-> BVar s (Tensor '[inPlane, oW, oH])

run a backprop-aware dilatedMaxPooling2d function

dilatedMaxPooling2dBatch

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH dilH dilW ceilMode

=> KnownDim inPlane

=> KnownDim b

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> Dilation2d dilH dilW
dilation size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[b, inPlane, iW, iH])

-> BVar s (Tensor '[b, inPlane, oW, oH])

run a backprop-aware dilatedMaxPooling2d function with a batch dimension.

_dilatedMaxPooling2d

Arguments

:: All KnownDim '[kH, kW, pH, pW, dH, dW, dilH, dilW]

=> All Dimensions '[d', d]

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> Dilation2d dilH dilW
dilation size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor d)
input

-> BVar s (Tensor d')
output

internal function of dilatedMaxPooling2d and dilatedMaxPooling2dBatch. Should not be used.

2d max pooling helpers

_maxPooling2d

Arguments

:: All KnownDim '[kH, kW, pH, pW, dH, dW]

=> All Dimensions '[d', d]

=> Reifies s W

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size. Note: default in C is the kernel size.

-> Padding2d pH pW
padding size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor d)
input

-> BVar s (Tensor d')
output

internal function of maxPooling2d and maxPooling2dBatch. Should not be used.

maxPooling2d

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH 1 1 ceilMode

=> Reifies s W

=> KnownDim inPlane

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[inPlane, iH, iW])

-> BVar s (Tensor '[inPlane, oH, oW])

backprop-aware maxPooling2d function.

maxPooling2dBatch

Arguments

:: SpatialDilationC iH iW kH kW dH dW pH pW oW oH 1 1 ceilMode

=> Reifies s W

=> KnownDim inPlane

=> KnownDim b

=> Kernel2d kH kW
kernel size

-> Step2d dH dW
step size

-> Padding2d pH pW
padding size

-> SBool ceilMode
ceil mode

-> BVar s (Tensor '[b, inPlane, iH, iW])

-> BVar s (Tensor '[b, inPlane, oH, oW])

backprop-aware maxPooling2d function with a batch dimension.

_spatialAdaptiveMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

spatialAdaptiveMaxPooling forward pass (updates the output tensor)

_spatialAdaptiveMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> IO ()

spatialAdaptiveMaxPooling backward-update (updates the layer and bias tensors)

_spatialFractionalMaxPooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IndexTensor d -> Tensor d -> IO ()

spatialFractionalMaxPooling forward pass (updates the output tensor)

_spatialFractionalMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> IndexTensor d -> IO ()

spatialFractionalMaxPooling backward-update (updates the layer and bias tensors)

_spatialMaxUnpooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

spatialMaxUnpooling forward pass (updates the output tensor)

_spatialMaxUnpooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> IO ()

spatialMaxUnpooling backward-update (updates the layer and bias tensors)

_spatialAdaptiveAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> IO ()

spatialAdaptiveAveragePooling forward pass (updates the output tensor)

_spatialAdaptiveAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

spatialAdaptiveAveragePooling backward-update (updates the layer and bias tensors)

_spatialAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

spatialAveragePooling forward pass (updates the output tensor)

_spatialAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

spatialAveragePooling backward-update (updates the layer and bias tensors)

3D pooling functions

_volumetricFractionalMaxPooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IndexTensor d -> Tensor d -> IO ()

volumetricFractionalMaxPooling forward pass (updates the output tensor)

_volumetricFractionalMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> IndexTensor d -> IO ()

volumetricFractionalMaxPooling backward-update (updates the layer and bias tensors)

_volumetricMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricMaxPooling forward pass (updates the output tensor)

_volumetricMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricMaxPooling backward-update (updates the layer and bias tensors)

_volumetricDilatedMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricDilatedMaxPooling forward pass (updates the output tensor)

_volumetricDilatedMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> IO ()

volumetricDilatedMaxPooling backward-update (updates the layer and bias tensors)

_volumetricMaxUnpooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricMaxUnpooling forward pass (updates the output tensor)

_volumetricMaxUnpooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> IO ()

volumetricMaxUnpooling backward-update (updates the layer and bias tensors)

_volumetricAdaptiveMaxPooling_updateOutput :: Tensor d -> Tensor d -> IndexTensor d -> Int -> Int -> Int -> IO ()

volumetricAdaptiveMaxPooling forward pass (updates the output tensor)

_volumetricAdaptiveMaxPooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IndexTensor d -> IO ()

volumetricAdaptiveMaxPooling backward-update (updates the layer and bias tensors)

_volumetricAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

volumetricAveragePooling forward pass (updates the output tensor)

_volumetricAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Bool -> Bool -> IO ()

volumetricAveragePooling backward-update (updates the layer and bias tensors)

_volumetricAdaptiveAveragePooling_updateOutput :: Tensor d -> Tensor d -> Int -> Int -> Int -> IO ()

volumetricAdaptiveAveragePooling forward pass (updates the output tensor)

_volumetricAdaptiveAveragePooling_updateGradInput :: Tensor d -> Tensor d -> Tensor d -> IO ()

volumetricAdaptiveAveragePooling backward-update (updates the layer and bias tensors)