Torch.Indef.Static.Tensor

Description

Documentation

vector :: forall n. KnownDim n => KnownNat n => [HsReal] -> Maybe (Tensor '[n])

Purely make a 1d tensor from a list of unknown length.

newExpand :: Tensor d1 -> IndexStorage -> IO (Tensor d2)

Static call to newExpand

_expand :: Tensor d1 -> Tensor d2 -> IndexStorage -> IO ()

Static call to _expand

_expandNd :: NonEmpty (Tensor d1) -> NonEmpty (Tensor d2) -> Int -> IO ()

Static call to _expandNd

_resize :: Tensor d1 -> IndexStorage -> IndexStorage -> IO (Tensor d2)

Static call to _resize

_resize1d :: Tensor d1 -> Int64 -> IO (Tensor d2)

Static call to _resize1d

_resize2d :: Tensor d1 -> Int64 -> Int64 -> IO (Tensor d2)

Static call to _resize2d

_resize3d :: Tensor d1 -> Int64 -> Int64 -> Int64 -> IO (Tensor d2)

Static call to _resize3d

_resize4d :: Tensor d1 -> Int64 -> Int64 -> Int64 -> Int64 -> IO (Tensor d2)

Static call to _resize4d

_resize5d :: Tensor d1 -> Int64 -> Int64 -> Int64 -> Int64 -> Int64 -> IO (Tensor d2)

Static call to _resize5d

_resizeAs :: Tensor d1 -> Tensor d2 -> IO (Tensor d3)

Static call to _resizeAs

_resizeNd :: Tensor d1 -> Int32 -> [Size] -> [Stride] -> IO (Tensor d2)

Static call to _resizeNd

retain :: Tensor d -> IO ()

Static call to retain

_clearFlag :: Tensor d -> Int8 -> IO ()

Static call to _clearFlag

tensordata :: Tensor d -> IO [HsReal]

Static call to tensordata

get1d :: Tensor d -> Int64 -> HsReal

Static call to get1d

get2d :: Tensor d -> Int64 -> Int64 -> HsReal

Static call to get2d

get3d :: Tensor d -> Int64 -> Int64 -> Int64 -> HsReal

Static call to get3d

get4d :: Tensor d -> Int64 -> Int64 -> Int64 -> Int64 -> HsReal

Static call to get4d

isContiguous :: Tensor d -> IO Bool

Static call to isContiguous

isSetTo :: Tensor d1 -> Tensor d2 -> IO Bool

Static call to isSetTo

isSize :: Tensor d -> LongStorage -> IO Bool

Static call to isSize

nDimension :: Tensor d -> IO Int32

Static call to nDimension

nElement :: Tensor d -> IO Int64

Static call to nElement

_narrow :: Tensor d1 -> Tensor d2 -> DimVal -> Int64 -> Size -> IO ()

Static call to _narrow

empty :: IO (Tensor d)

renamed from TH's new because this always returns an empty tensor FIXME: this __technically should be IO (Tensor '[]), but if you leave it as-is the types line-up nicely (and we currently don't use rank-0 tensors).

newClone :: Tensor d1 -> IO (Tensor d2)

Static call to newClone

newContiguous :: Tensor d1 -> IO (Tensor d2)

Static call to newContiguous

newNarrow :: Tensor d1 -> DimVal -> Int64 -> Size -> IO (Tensor d2)

Static call to newNarrow

newSelect :: Tensor d1 -> DimVal -> Int64 -> IO (Tensor d2)

Static call to newSelect

newSizeOf :: Tensor d -> IO IndexStorage

Static call to newSizeOf

newStrideOf :: Tensor d -> IO IndexStorage

Static call to newStrideOf

newTranspose :: Tensor d1 -> DimVal -> DimVal -> IO (Tensor d2)

Static call to newTranspose

newUnfold :: Tensor d1 -> DimVal -> Int64 -> Int64 -> IO (Tensor d2)

Static call to newUnfold

view :: forall d d'. (Dimensions d, Dimensions d') => Tensor d -> IO (Tensor d')

Make a new view of a tensor.

newWithSize :: IndexStorage -> IndexStorage -> IO (Tensor d)

Static call to newWithSize

newWithSize1d :: Size -> IO (Tensor d)

Static call to newWithSize1d

newWithSize2d :: Size -> Size -> IO (Tensor d)

Static call to newWithSize2d

newWithSize3d :: Size -> Size -> Size -> IO (Tensor d)

Static call to newWithSize3d

newWithSize4d :: Size -> Size -> Size -> Size -> IO (Tensor d)

Static call to newWithSize4d

newWithStorage :: Storage -> StorageOffset -> IndexStorage -> IndexStorage -> IO (Tensor d)

Static call to newWithStorage

newWithStorage1d :: Storage -> StorageOffset -> (Size, Stride) -> IO (Tensor d)

Static call to newWithStorage1d

newWithStorage2d :: Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> IO (Tensor d)

Static call to newWithStorage2d

newWithStorage3d :: Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO (Tensor d)

Static call to newWithStorage3d

newWithStorage4d :: Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO (Tensor d)

Static call to newWithStorage4d

newWithTensor :: Tensor d1 -> IO (Tensor d2)

Static call to newWithTensor

_select :: Tensor d1 -> Tensor d2 -> DimVal -> Int64 -> IO ()

Static call to _select

_set :: Tensor d1 -> Tensor d2 -> IO ()

Static call to _set

_set1d :: Tensor d -> Int64 -> HsReal -> IO ()

Static call to _set1d

_set2d :: Tensor d -> Int64 -> Int64 -> HsReal -> IO ()

Static call to _set2d

_set3d :: Tensor d -> Int64 -> Int64 -> Int64 -> HsReal -> IO ()

Static call to _set3d

_set4d :: Tensor d -> Int64 -> Int64 -> Int64 -> Int64 -> HsReal -> IO ()

Static call to _set4d

_setFlag :: Tensor d -> Int8 -> IO ()

Static call to _setFlag

_setStorage :: Tensor d -> Storage -> StorageOffset -> IndexStorage -> IndexStorage -> IO ()

Static call to _setStorage

_setStorage1d :: Tensor d -> Storage -> StorageOffset -> (Size, Stride) -> IO ()

Static call to _setStorage1d

_setStorage2d :: Tensor d -> Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> IO ()

Static call to _setStorage2d

_setStorage3d :: Tensor d -> Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO ()

Static call to _setStorage3d

_setStorage4d :: Tensor d -> Storage -> StorageOffset -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> (Size, Stride) -> IO ()

Static call to _setStorage4d

_setStorageNd :: Tensor d -> Storage -> StorageOffset -> DimVal -> [Size] -> [Stride] -> IO ()

Static call to _setStorageNd

size :: Tensor d -> DimVal -> IO Size

Static call to size

sizeDesc :: Tensor d -> IO DescBuff

Static call to sizeDesc

_squeeze :: Tensor d1 -> Tensor d2 -> IO ()

Static call to _squeeze

squeeze1d :: Dimensions d => Dim n -> Tensor d -> Tensor d'

Squeeze a dimension of size 1 out of the tensor

storage :: Tensor d -> IO Storage

Static call to storage

storageOffset :: Tensor d -> IO StorageOffset

Static call to storageOffset

stride :: Tensor d -> DimVal -> IO Stride

Static call to stride

_transpose :: Tensor d1 -> Tensor d2 -> DimVal -> DimVal -> IO ()

Static call to _transpose

_unfold :: Tensor d1 -> Tensor d2 -> DimVal -> Size -> Step -> IO ()

Static call to _unfold

unsqueeze1d :: All Dimensions '[d, d'] => Idx n -> Tensor d -> Tensor d'

Unsqueeze a dimension of size 1 into the tensor (pure, dupable)

unsqueeze1d_ :: All Dimensions '[d, d'] => Idx n -> Tensor d -> IO (Tensor d')

Not safe:* unsqueeze a dimension of size 1 into the tensor.

shape :: Tensor d -> IO [Size]

Get runtime shape information from a tensor

getDims :: Tensor d -> IO SomeDims

alias to shape, casting the dimensions into a runtime SomeDims.

withNew :: forall d. Dimensions d => (Tensor d -> IO ()) -> IO (Tensor d)

helper function to make other parts of hasktorch valid pure functions.

withEmpty :: forall d. Dimensions d => (Tensor d -> IO ()) -> IO (Tensor d)

Should be renamed to newFromSize

withEmpty' :: Dimensions d => (Tensor d -> IO ()) -> IO (Tensor d)

same as withEmpty (which should be called newFromSize) and withNew, but passes in an empty tensor to be mutated and returned with a static dimensionality that it is assumed to take on after the mutation.

Note: We can get away with this when Torch does resizing in C, but you need to examine the C implementation of the function you are trying to make pure.

withInplace :: Dimensions d => Tensor d -> (Tensor d -> Tensor d -> IO ()) -> IO (Tensor d)

Deprecated: this is a trivial function with a bad API

This is actually inplace. Dimensions may change from original tensor given Torch resizing.

FIXME: remove this function

throwFIXME :: MonadThrow io => String -> String -> io x

throw a FIXME string.

throwNE :: MonadThrow io => String -> io x

throw an "unsafe head" string.

throwGT4 :: MonadThrow io => String -> io x

throw an "unsupported dimension" string.

_setStorageDim :: Tensor d -> Storage -> StorageOffset -> [(Size, Stride)] -> IO ()

Set the storage of a tensor. This is incredibly unsafe.

setDim_ :: Tensor d -> Dims (d' :: [Nat]) -> HsReal -> IO ()

Set the value of a tensor at a given index

FIXME: there should be a constraint to see that d' is in d

setDim'_ :: Tensor d -> SomeDims -> HsReal -> IO ()

runtime version of setDim_

getDim :: forall d d'. All Dimensions '[d, d'] => Tensor (d :: [Nat]) -> Dims (d' :: [Nat]) -> IO HsReal

get the value of a tensor at the given index

FIXME: there should be a constraint to see that d' is in d

(!!) :: forall (d :: [Nat]) (d' :: [Nat]). All Dimensions '[d, d'] => Tensor d -> DimVal -> Tensor d'

Select a dimension of a tensor. If a vector is passed in, return a singleton tensor with the index value of the vector.

new :: forall d. Dimensions d => IO (Tensor d)

Create a new tensor. Elements have not yet been allocated and there will not be any gauruntees regarding what is inside.

_resizeDim :: forall d d'. Dimensions d' => Tensor d -> IO (Tensor d')

Resize a tensor, returning the same tensor with a the changed dimensions.

NOTE: This is copied from the dynamic version to keep the constraints clean and is _unsafe_

resizeAs_ :: forall d d'. (All Dimensions '[d, d'], Product d ~ Product d') => Tensor d -> IO (Tensor d')

Resize the input with the output shape. impure and mutates the tensor inplace.

FIXME: replace d with a linear type?

resizeAs :: forall d d'. (All Dimensions [d, d'], Product d ~ Product d') => Tensor d -> Tensor d'

Pure version of resizeAs_ which clones the input tensor (pure, dupable?)

flatten :: (Dimensions d, KnownDim (Product d)) => Tensor d -> Tensor '[Product d]

flatten a tensor (pure, dupable)

fromList :: forall d. Dimensions d => KnownNat (Product d) => KnownDim (Product d) => [HsReal] -> Maybe (Tensor d)

Initialize a tensor of arbitrary dimension from a list FIXME: There might be a faster way to do this with newWithSize

matrix :: forall n m. All KnownDim '[n, m] => All KnownNat '[n, m] => KnownDim (n * m) => KnownNat (n * m) => [[HsReal]] -> Either String (Tensor '[n, m])

Purely construct a matrix from a list of lists. This assumes that the list of lists is a dense matrix representation. Returns either the successful construction of the tensor, or a string explaining what went wrong.

transpose2d :: All KnownDim '[r, c] => Tensor '[r, c] -> Tensor '[c, r]

transpose a matrix (pure, dupable)

expand2d :: forall x y. All KnownDim '[x, y] => Tensor '[x] -> IO (Tensor '[y, x])

Expand a vector by copying into a matrix by set dimensions TODO - generalize this beyond the matrix case

getElem2d :: forall (n :: Nat) (m :: Nat). All KnownDim '[n, m] => Tensor '[n, m] -> Natural -> Natural -> IO HsReal

Deprecated: use getDim instead

Get an element from a matrix with runtime index values.

FIXME: This is primarily for backwards compatability with lasso and should be removed.

setElem2d :: forall (n :: Nat) (m :: Nat) ns. All KnownDim '[n, m] => Tensor '[n, m] -> Natural -> Natural -> HsReal -> IO ()

Deprecated: use setDim_ instead

Set an element on a matrix with runtime index values.

FIXME: This is primarily for backwards compatability with lasso and should be removed.

Orphan instances

Dimensions d => Show (Tensor d)

Instance details

Methods

showsPrec :: Int -> Tensor d -> ShowS

show :: Tensor d -> String

showList :: [Tensor d] -> ShowS