Torch.Indef.Static.Tensor.Math

Description

Documentation

fill_ :: Tensor d -> HsReal -> IO ()

Static call to fill_

zero_ :: Tensor d -> IO ()

Static call to zero_

zeros_ :: Tensor d -> IndexStorage -> IO ()

mutate a tensor, inplace, resizing the tensor to the given IndexStorage size and replacing its value with zeros.

zerosLike_

Arguments

:: Tensor d
tensor to mutate inplace and replace contents with zeros

-> Tensor d'
tensor to extract shape information from.

-> IO ()

mutate a tensor, inplace, resizing the tensor to the same shape as the second tensor argument and replacing the first tensor's values with zeros.

ones_ :: Tensor d -> IndexStorage -> IO ()

mutate a tensor, inplace, resizing the tensor to the given IndexStorage size and replacing its value with ones.

onesLike_

Arguments

:: Tensor d
tensor to mutate inplace and replace contents with ones

-> Tensor d'
tensor to extract shape information from.

-> IO ()

mutate a tensor, inplace, resizing the tensor to the same shape as the second tensor argument and replacing the first tensor's values with ones.

numel :: Tensor d -> IO Integer

Static call to numel

_reshape :: Tensor d1 -> Tensor d2 -> IndexStorage -> IO ()

Static call to _reshape

_catArray :: Tensor d -> [Dynamic] -> DimVal -> IO ()

Static call to _catArray

_nonzero :: LongTensor d1 -> Tensor d2 -> IO ()

Static call to _nonzero

_tril :: Tensor d1 -> Tensor d2 -> Integer -> IO ()

Static call to _tril

_triu :: Tensor d1 -> Tensor d2 -> Integer -> IO ()

Static call to _triu

eye_ :: Tensor d -> Integer -> Integer -> IO ()

Static call to eye_

trace :: Tensor d -> IO HsAccReal

Static call to trace

_arange :: Tensor d -> HsAccReal -> HsAccReal -> HsAccReal -> IO ()

Static call to _arange

range_ :: Tensor d -> HsAccReal -> HsAccReal -> HsAccReal -> IO ()

Static call to range_

constant :: forall d. Dimensions d => HsReal -> Tensor d

Static call to constant

diag_ :: All Dimensions '[d, d'] => Tensor d -> Int -> IO (Tensor d')

Static call to diag_

diag :: All Dimensions '[d, d'] => Tensor d -> Int -> IO (Tensor d')

Static call to diag

diag1d :: KnownDim n => Tensor '[n] -> IO (Tensor '[n, n])

Create a diagonal matrix from a 1D vector

cat_ :: All Dimensions '[d, d', d''] => Tensor d -> Tensor d' -> DimVal -> IO (Tensor d'')

Warning: this function is impure and the dimensions can fall out of sync with the type, if used incorrectly

Static call to cat_. Unsafely returning the resulting tensor with new dimensions.

cat :: All Dimensions '[d, d', d''] => Tensor d -> Tensor d' -> DimVal -> IO (Tensor d'')

Static call to cat

cat1d :: (All KnownDim '[n1, n2, n], n ~ Sum [n1, n2]) => Tensor '[n1] -> Tensor '[n2] -> IO (Tensor '[n])

convenience function, specifying a type-safe cat operation.

cat2d1 :: (All KnownDim '[n, m, m0, m1], m ~ Sum [m0, m1]) => Tensor '[n, m0] -> Tensor '[n, m1] -> IO (Tensor '[n, m])

convenience function, specifying a type-safe cat operation.

cat2d0 :: (All KnownDim '[n, m, n0, n1], n ~ Sum [n0, n1]) => Tensor '[n0, m] -> Tensor '[n1, m] -> IO (Tensor '[n, m])

convenience function, specifying a type-safe cat operation.

cat3d0 :: (All KnownDim '[x, y, x0, x1, z], x ~ Sum [x0, x1]) => Tensor '[x0, y, z] -> Tensor '[x1, y, z] -> IO (Tensor '[x, y, z])

convenience function, specifying a type-safe cat operation.

cat3d1 :: (All KnownDim '[x, y, y0, y1, z], y ~ Sum [y0, y1]) => Tensor '[x, y0, z] -> Tensor '[x, y1, z] -> IO (Tensor '[x, y, z])

convenience function, specifying a type-safe cat operation.

cat3d2 :: (All KnownDim '[x, y, z0, z1, z], z ~ Sum [z0, z1]) => Tensor '[x, y, z0] -> Tensor '[x, y, z1] -> IO (Tensor '[x, y, z])

convenience function, specifying a type-safe cat operation.

catArray :: Dimensions d => [Dynamic] -> DimVal -> IO (Tensor d)

Concatenate all tensors in a given list of dynamic tensors along the given dimension.

NOTE: In C, if the dimension is not specified or if it is -1, it is the maximum last dimension over all input tensors, except if all tensors are empty, then it is 1.

C-Style: In the classic Torch C-style, the first argument is treated as the return type and is mutated in-place.

onesLike :: forall d. Dimensions d => IO (Tensor d)

Static call to onesLike

zerosLike :: forall d. Dimensions d => IO (Tensor d)

Static call to zerosLike