Torch.Indef.Static.Tensor.Math.CompareT

Description

Documentation

ltTensor :: Dimensions d => Tensor d -> Tensor d -> MaskTensor d

Return a byte tensor which contains boolean values indicating the relation between two tensors.

ltTensorT

Arguments

:: Dimensions d

=> Tensor d
source tensor.

-> Tensor d
tensor to compare with.

-> Tensor d
new return tensor.

return a tensor which contains numeric values indicating the relation between two tensors. 0 stands for false, 1 stands for true.

ltTensorT_

Arguments

:: Tensor d
source tensor to mutate inplace.

-> Tensor d
tensor to compare with.

-> IO ()

mutate a tensor in-place with its numeric relation to the second tensor of the same size, where 0 stands for false and 1 stands for true.

leTensor :: Dimensions d => Tensor d -> Tensor d -> MaskTensor d

Return a byte tensor which contains boolean values indicating the relation between two tensors.

leTensorT

Arguments

:: Dimensions d

=> Tensor d
source tensor.

-> Tensor d
tensor to compare with.

-> Tensor d
new return tensor.

return a tensor which contains numeric values indicating the relation between two tensors. 0 stands for false, 1 stands for true.

leTensorT_

Arguments

:: Tensor d
source tensor to mutate inplace.

-> Tensor d
tensor to compare with.

-> IO ()

mutate a tensor in-place with its numeric relation to the second tensor of the same size, where 0 stands for false and 1 stands for true.

gtTensor :: Dimensions d => Tensor d -> Tensor d -> MaskTensor d

Return a byte tensor which contains boolean values indicating the relation between two tensors.

gtTensorT

Arguments

:: Dimensions d

=> Tensor d
source tensor.

-> Tensor d
tensor to compare with.

-> Tensor d
new return tensor.

return a tensor which contains numeric values indicating the relation between two tensors. 0 stands for false, 1 stands for true.

gtTensorT_

Arguments

:: Tensor d
source tensor to mutate inplace.

-> Tensor d
tensor to compare with.

-> IO ()

mutate a tensor in-place with its numeric relation to the second tensor of the same size, where 0 stands for false and 1 stands for true.

geTensor :: Dimensions d => Tensor d -> Tensor d -> MaskTensor d

Return a byte tensor which contains boolean values indicating the relation between two tensors.

geTensorT

Arguments

:: Dimensions d

=> Tensor d
source tensor.

-> Tensor d
tensor to compare with.

-> Tensor d
new return tensor.

return a tensor which contains numeric values indicating the relation between two tensors. 0 stands for false, 1 stands for true.

geTensorT_

Arguments

:: Tensor d
source tensor to mutate inplace.

-> Tensor d
tensor to compare with.

-> IO ()

mutate a tensor in-place with its numeric relation to the second tensor of the same size, where 0 stands for false and 1 stands for true.

neTensor :: Dimensions d => Tensor d -> Tensor d -> MaskTensor d

Return a byte tensor which contains boolean values indicating the relation between two tensors.

neTensorT

Arguments

:: Dimensions d

=> Tensor d
source tensor.

-> Tensor d
tensor to compare with.

-> Tensor d
new return tensor.

return a tensor which contains numeric values indicating the relation between two tensors. 0 stands for false, 1 stands for true.

neTensorT_

Arguments

:: Tensor d
source tensor to mutate inplace.

-> Tensor d
tensor to compare with.

-> IO ()

mutate a tensor in-place with its numeric relation to the second tensor of the same size, where 0 stands for false and 1 stands for true.

eqTensor :: Dimensions d => Tensor d -> Tensor d -> MaskTensor d

Return a byte tensor which contains boolean values indicating the relation between two tensors.

eqTensorT

Arguments

:: Dimensions d

=> Tensor d
source tensor.

-> Tensor d
tensor to compare with.

-> Tensor d
new return tensor.

return a tensor which contains numeric values indicating the relation between two tensors. 0 stands for false, 1 stands for true.

eqTensorT_

Arguments

:: Tensor d
source tensor to mutate inplace.

-> Tensor d
tensor to compare with.

-> IO ()

mutate a tensor in-place with its numeric relation to the second tensor of the same size, where 0 stands for false and 1 stands for true.