Torch.Indef.Static.Tensor.Random.TH

Description

Documentation

_random :: Dimensions d => Tensor d -> Generator -> IO ()

Static call to _random

random :: forall d. Dimensions d => Generator -> IO (Tensor d)

Static call to random

_clampedRandom :: Dimensions d => Tensor d -> Generator -> Ord2Tuple Integer -> IO ()

Static call to _clampedRandom

clampedRandom :: forall d. Dimensions d => Generator -> Ord2Tuple Integer -> IO (Tensor d)

Static call to clampedRandom

_cappedRandom :: Dimensions d => Tensor d -> Generator -> Integer -> IO ()

Static call to _cappedRandom

cappedRandom :: forall d. Dimensions d => Generator -> Word64 -> IO (Tensor d)

Static call to cappedRandom

_geometric :: Dimensions d => Tensor d -> Generator -> HsAccReal -> IO ()

Static call to _geometric

geometric :: forall d. Dimensions d => Generator -> OpenUnit HsAccReal -> IO (Tensor d)

Static call to geometric

_bernoulli :: Dimensions d => Tensor d -> Generator -> HsAccReal -> IO ()

Static call to _bernoulli

bernoulli :: forall d. Dimensions d => Generator -> ClosedUnit HsAccReal -> IO (Tensor d)

Static call to bernoulli

_bernoulli_FloatTensor :: Dimensions d => Tensor d -> Generator -> FloatTensor d -> IO ()

Static call to _bernoulli_FloatTensor

bernoulli_FloatTensor :: forall d. Dimensions d => Generator -> FloatTensor d -> IO (Tensor d)

Static call to bernoulli_FloatTensor

_bernoulli_DoubleTensor :: Dimensions d => Tensor d -> Generator -> DoubleTensor d -> IO ()

Static call to _bernoulli_DoubleTensor

bernoulli_DoubleTensor :: forall d. Dimensions d => Generator -> DoubleTensor d -> IO (Tensor d)

Static call to bernoulli_DoubleTensor

_uniform :: Dimensions d => Tensor d -> Generator -> Ord2Tuple HsAccReal -> IO ()

Static call to _uniform

uniform :: forall d. Dimensions d => Generator -> Ord2Tuple HsAccReal -> IO (Tensor d)

Static call to uniform

_normal :: Dimensions d => Tensor d -> Generator -> HsAccReal -> Positive HsAccReal -> IO ()

Static call to _normal

normal :: forall d. Dimensions d => Generator -> HsAccReal -> Positive HsAccReal -> IO (Tensor d)

Static call to normal

_normal_means :: Dimensions d => Tensor d -> Generator -> Tensor d -> Positive HsAccReal -> IO ()

Static call to _normal_means

normal_means :: forall d. Dimensions d => Generator -> Tensor d -> Positive HsAccReal -> IO (Tensor d)

Static call to normal_means

_normal_stddevs :: Dimensions d => Tensor d -> Generator -> HsAccReal -> Tensor d -> IO ()

Static call to _normal_stddevs

normal_stddevs :: forall d. Dimensions d => Generator -> HsAccReal -> Tensor d -> IO (Tensor d)

Static call to normal_stddevs

_normal_means_stddevs :: Dimensions d => Tensor d -> Generator -> Tensor d -> Tensor d -> IO ()

Static call to _normal_means_stddevs

normal_means_stddevs :: forall d. Dimensions d => Generator -> Tensor d -> Tensor d -> IO (Tensor d)

Static call to normal_means_stddevs

_exponential :: Dimensions d => Tensor d -> Generator -> HsAccReal -> IO ()

Static call to _exponential

exponential :: forall d. Dimensions d => Generator -> HsAccReal -> IO (Tensor d)

Static call to exponential

_standard_gamma :: Dimensions d => Tensor d -> Generator -> Tensor d -> IO ()

Static call to _standard_gamma

standard_gamma :: forall d. Dimensions d => Generator -> Tensor d -> IO (Tensor d)

Static call to standard_gamma

_cauchy :: Dimensions d => Tensor d -> Generator -> HsAccReal -> HsAccReal -> IO ()

Static call to _cauchy

cauchy :: forall d. Dimensions d => Generator -> HsAccReal -> HsAccReal -> IO (Tensor d)

Static call to cauchy

_logNormal :: Dimensions d => Tensor d -> Generator -> HsAccReal -> Positive HsAccReal -> IO ()

Static call to _logNormal

logNormal :: forall d. Dimensions d => Generator -> HsAccReal -> Positive HsAccReal -> IO (Tensor d)

Static call to logNormal

_multinomial :: Dimensions d => IndexTensor d -> Generator -> Tensor d -> Int -> Int -> IO ()

Static call to _multinomial

_multinomialAliasSetup :: Dimensions d => Tensor d -> IndexTensor d -> Tensor d -> IO ()

Static call to _multinomialAliasSetup

_multinomialAliasDraw :: Dimensions d => IndexTensor d -> Generator -> IndexTensor d -> Tensor d -> IO ()

Static call to _multinomialAliasDraw

data OpenUnit x

Datatype to represent the open unit interval: 0 < x < 1. Any OpenUnit inhabitant must satisfy being in the interval.

FIXME: replace with numhask.

openUnit :: (Ord x, Num x) => x -> Maybe (OpenUnit x)

smart constructor to place a number in the open unit interval.

openUnitValue :: OpenUnit x -> x

Get a value from the open unit interval.

data ClosedUnit x

Datatype to represent the closed unit interval: 0 =< x =< 1. Any ClosedUnit inhabitant must satisfy being in the interval.

FIXME: replace with numhask.

closedUnit :: (Ord x, Num x) => x -> Maybe (ClosedUnit x)

smart constructor to place a number in the closed unit interval.

closedUnitValue :: ClosedUnit x -> x

Get a value from the closed unit interval.

data Positive x

Datatype to represent a generic positive number: 0 =< x.

FIXME: replace with numhask.

positive :: (Ord x, Num x) => x -> Maybe (Positive x)

smart constructor to place a number in a positive bound.

positiveValue :: Positive x -> x

Get a value from the positive bound.

data Ord2Tuple x

Datatype to represent an ordered pair of numbers, (a, b), where a <= b.

FIXME: replace with numhask.

ord2Tuple :: (Ord x, Num x) => (x, x) -> Maybe (Ord2Tuple x)

smart constructor to place two values in an ordered tuple.

ord2TupleValues :: Ord2Tuple x -> (x, x)

Get the values of an ordered tuple.

multivariate_normal

Arguments

:: All KnownDim '[n, p]

=> Generator

-> Tensor '[p]
mu

-> Tensor '[p, p]
eigenvec

-> Tensor '[p]
eigenval

-> IO (Tensor '[n, p])

find the multivariate normal distribution given mu, an eigenvector and eigenvalues.