Torch.Indef.Storage

Description

Storages are basically a way to access memory of a C pointer or array. Storages can also map the contents of a file to memory. A Storage is an array of basic C types.

Several Storage classes for all the basic C types exist and have the following self-explanatory names: ByteStorage, CharStorage, ShortStorage, IntStorage, LongStorage, FloatStorage, DoubleStorage.

Note that ByteStorage and CharStorage represent both arrays of bytes. ByteStorage represents an array of unsigned chars, while CharStorage represents an array of signed chars.

Documentation

module Torch.Indef.Storage.Copy

tensordata :: Storage -> IO [HsReal]

return the internal data of Storage as a list of haskell values.

size :: Storage -> IO StorageSize

Returns the number of elements in the storage. Equivalent to #.

set :: Storage -> Index -> HsReal -> IO ()

set the value at Index to HsReal in a given Storage.

get :: Storage -> Index -> IO HsReal

get the value at Index from a given Storage.

empty :: IO Storage

make a new empty Storage.

newWithSize :: StorageSize -> IO Storage

create a new storage of a given length, StorageSize.

newWithSize1 :: HsReal -> IO Storage

make a new Storage with a single value.

newWithSize2 :: HsReal -> HsReal -> IO Storage

make a new Storage with two values.

newWithSize3 :: HsReal -> HsReal -> HsReal -> IO Storage

make a new Storage with three values.

newWithSize4 :: HsReal -> HsReal -> HsReal -> HsReal -> IO Storage

make a new Storage with four values.

newWithMapping

Arguments

:: [Int8]
filename

-> StorageSize
size

-> Int32
flags

-> IO Storage

FIXME: This is totally broken. This takes a filename, size, and flags, and produces Storage from these inputs. Figure out how to fix this, ideally.

See: https://github.com/torch/torch7/blob/04e1d1dce0f02aea82dc433c4f39e42650c4390f/lib/TH/generic/THStorage.h#L49

newWithData :: [HsReal] -> StorageSize -> IO Storage

make a new Storage from a given list and StorageSize.

FIXME: find out if StorageSize always corresponds to the length of the list. If so, remove it!

setFlag :: Storage -> Int8 -> IO ()

set the flags of a given Storage. Flags are applied via bitwise-or.

clearFlag :: Storage -> Int8 -> IO ()

clear the flags of a given Storage. Flags are cleanred via bitwise-and.

retain :: Storage -> IO ()

Increment the reference counter of the storage.

This method should be used with extreme care. In general, they should never be called, except if you know what you are doing, as the handling of references is done automatically. They can be useful in threaded environments. Note that these methods are atomic operations.

resize :: Storage -> StorageSize -> IO ()

Resize the storage to the provided size. The new contents are undetermined.

fill :: Storage -> HsReal -> IO ()

Fill the Storage with the given value.