Torch.Indef.Types

Description

Documentation

module Torch.Sig.Types.Global

module Torch.Sig.Types

module Torch.Sig.State

class THDebug t where

Class to print out C-references when working debugging segfaults.

Minimal complete definition

printRefs

Methods

printRefs :: t -> IO ()

print out all possible references we can find.

newtype DimVal

Deprecated: Use dimensions package's Idx instead

term-level representation of an index.

Constructors

DimVal Int32

Deprecated: Use dimensions package's Idx instead

newtype Step

newtype wrapper around the C-level representation of a step size

Constructors

Step CLong

newtype Stride

newtype wrapper around the C-level representation of a tensor's internal Storage stride for each dimension.

Constructors

Stride CLLong

newtype StorageOffset

newtype wrapper around the C-level representation of a storage offset

Constructors

StorageOffset CPtrdiff

newtype Size

newtype wrapper around the C-level representation of a dimension's size

Constructors

Size CLLong

newtype KeepDim

haskell representation of a CInt which determines whether or not to return dimensions

Constructors

KeepDim

Fields

keepIt :: Bool

fromKeepDim :: Integral i => Maybe KeepDim -> i

cast a KeepDim to a numerical representation.

NOTE: don't bind the i in case there are some differences between THC and TH

keep :: KeepDim

smart constructors for keepdim since we don't get inference for free like Num

ignore :: KeepDim

smart constructors for keepdim since we don't get inference for free like Num

data SortOrder

Simple datatype to represent sort order arguments which torch provides to us.

Constructors

Ascending

Descending

data TopKOrder

Simple datatype to represent arguments for a topk function.

See https://github.com/torch/torch7/blob/75a86469aa9e2f5f04e11895b269ec22eb0e4687/lib/TH/generic/THTensorMath.c#L2545

Constructors

KAscending

KNone

KDescending

newtype StorageSize

Represents the size of storage, should be CPtrdiff to match with the C internals

Constructors

StorageSize CPtrdiff

newtype AllocatorContext

Warning: this should not be used or referenced -- we are still figuring out what to do with this.

this is supposed to represent the AllocatorContext, but it should not be exposed to a user.

Constructors

AllocatorContext (Ptr ())

Warning: this should not be used or referenced -- we are still figuring out what to do with this.

newtype Index

Deprecated: Use dimensions package's Idx instead

term-level representation of an index.

Constructors

Index Int64

Deprecated: Use dimensions package's Idx instead

manage' :: (c -> ForeignPtr a) -> c -> Managed (Ptr a)

helper functions to start using the Managed Monad more.

FIXME: Try to replace a lot of the below with this function, but ultimately try to remove this helper.

joinIO :: MonadIO m => m (IO x) -> m x

helper function to join MonadIO and IO.

(.:) :: (b -> c) -> (a0 -> a1 -> b) -> a0 -> a1 -> c infixl 5

The blackbird combinator.

(stites): This happens often enough that I'm pulling in the blackbird

FIXME(stites): remove this

(..:) :: (b -> c) -> (a0 -> a1 -> a2 -> b) -> a0 -> a1 -> a2 -> c infixl 5

even more blackbird

FIXME(stites): remove this

shuffle2 :: (a -> b -> c -> d) -> c -> a -> b -> d

shuffle 2 arguments for pointfree raw-ffi functions.

FIXME(stites): remove this

shuffle2'2 :: (a -> b -> c -> d -> e) -> c -> d -> a -> b -> e

shuffle the first two arguments two positions to the right for pointfree raw-ffi functions.

FIXME(stites): remove this

shuffle3 :: (a -> b -> c -> d -> e) -> d -> a -> b -> c -> e

shuffle the first three arguments to the right for pointfree raw-ffi functions.

FIXME(stites): remove this

shuffle3'2 :: (a -> b -> c -> d -> e -> f) -> d -> e -> a -> b -> c -> f

shuffle the first three arguments two positions to the right for pointfree raw-ffi functions.

FIXME(stites): remove this

withGen :: Generator -> (Ptr CGenerator -> IO x) -> IO x

run a function against the internal reference of a torch generator.

withState :: State -> (Ptr CState -> IO x) -> IO x

run a function with a managed state's raw internal pointer.

withDynamicState :: Dynamic -> (Ptr CState -> Ptr CTensor -> IO x) -> IO x

run a function with access to a tensor's underlying state and C-tensor.

withStorageState :: Storage -> (Ptr CState -> Ptr CStorage -> IO x) -> IO x

run a function with access to a Storages underlying state and C-reference.

with2DynamicState :: Dynamic -> Dynamic -> (Ptr CState -> Ptr CTensor -> Ptr CTensor -> IO x) -> IO x

run a function with two tensors with reference to the first tensor's underlying state.

with3DynamicState :: Dynamic -> Dynamic -> Dynamic -> (Ptr CState -> Ptr CTensor -> Ptr CTensor -> Ptr CTensor -> IO x) -> IO x

run a function with three tensors with reference to the first tensor's underlying state.

mkDynamic :: Ptr CState -> Ptr CTensor -> IO Dynamic

smart constructor for a Dynamic tensor

mkStorage :: Ptr CState -> Ptr CStorage -> IO Storage

smart constructor for Storage

mkDynamicIO :: (Ptr CState -> IO (Ptr CTensor)) -> IO Dynamic

smart constructor for a Dynamic tensor with a given builder function.

mkStorageIO :: (Ptr CState -> IO (Ptr CStorage)) -> IO Storage

smart constructor for Storage with a given builder function.